@@ -8,8 +8,8 @@ try reading the *_test.go files.
     (c) 2013 COSEINC. All Rights Reserved.
 
     THIS FILE WAS AUTO-GENERATED -- DO NOT EDIT!
-	Command: ./genconst ../capstone/bindings/python/capstone/
-	2021-10-14T08:50:33-04:00
+	Command: ./genconst ./capstone/bindings/python/capstone
+	2025-03-26T14:48:26+00:00
 
 */
 
@@ -92,6 +92,17 @@ const (
 	ARM_SYSREG_BASEPRI_MAX  = C.ARM_SYSREG_BASEPRI_MAX
 	ARM_SYSREG_FAULTMASK    = C.ARM_SYSREG_FAULTMASK
 	ARM_SYSREG_CONTROL      = C.ARM_SYSREG_CONTROL
+	ARM_SYSREG_MSPLIM       = C.ARM_SYSREG_MSPLIM
+	ARM_SYSREG_PSPLIM       = C.ARM_SYSREG_PSPLIM
+	ARM_SYSREG_MSP_NS       = C.ARM_SYSREG_MSP_NS
+	ARM_SYSREG_PSP_NS       = C.ARM_SYSREG_PSP_NS
+	ARM_SYSREG_MSPLIM_NS    = C.ARM_SYSREG_MSPLIM_NS
+	ARM_SYSREG_PSPLIM_NS    = C.ARM_SYSREG_PSPLIM_NS
+	ARM_SYSREG_PRIMASK_NS   = C.ARM_SYSREG_PRIMASK_NS
+	ARM_SYSREG_BASEPRI_NS   = C.ARM_SYSREG_BASEPRI_NS
+	ARM_SYSREG_FAULTMASK_NS = C.ARM_SYSREG_FAULTMASK_NS
+	ARM_SYSREG_CONTROL_NS   = C.ARM_SYSREG_CONTROL_NS
+	ARM_SYSREG_SP_NS        = C.ARM_SYSREG_SP_NS
 	ARM_SYSREG_R8_USR       = C.ARM_SYSREG_R8_USR
 	ARM_SYSREG_R9_USR       = C.ARM_SYSREG_R9_USR
 	ARM_SYSREG_R10_USR      = C.ARM_SYSREG_R10_USR
@@ -194,6 +205,7 @@ const (
 	ARM_VECTORDATA_U32     = C.ARM_VECTORDATA_U32
 	ARM_VECTORDATA_U64     = C.ARM_VECTORDATA_U64
 	ARM_VECTORDATA_P8      = C.ARM_VECTORDATA_P8
+	ARM_VECTORDATA_F16     = C.ARM_VECTORDATA_F16
 	ARM_VECTORDATA_F32     = C.ARM_VECTORDATA_F32
 	ARM_VECTORDATA_F64     = C.ARM_VECTORDATA_F64
 	ARM_VECTORDATA_F16F64  = C.ARM_VECTORDATA_F16F64
@@ -218,6 +230,10 @@ const (
 	ARM_VECTORDATA_F64U16  = C.ARM_VECTORDATA_F64U16
 	ARM_VECTORDATA_F32U16  = C.ARM_VECTORDATA_F32U16
 	ARM_VECTORDATA_F64U32  = C.ARM_VECTORDATA_F64U32
+	ARM_VECTORDATA_F16U16  = C.ARM_VECTORDATA_F16U16
+	ARM_VECTORDATA_U16F16  = C.ARM_VECTORDATA_U16F16
+	ARM_VECTORDATA_F16U32  = C.ARM_VECTORDATA_F16U32
+	ARM_VECTORDATA_U32F16  = C.ARM_VECTORDATA_U32F16
 )
 
 const (
@@ -346,21 +362,27 @@ const (
 	ARM_INS_INVALID   = C.ARM_INS_INVALID
 	ARM_INS_ADC       = C.ARM_INS_ADC
 	ARM_INS_ADD       = C.ARM_INS_ADD
+	ARM_INS_ADDW      = C.ARM_INS_ADDW
 	ARM_INS_ADR       = C.ARM_INS_ADR
 	ARM_INS_AESD      = C.ARM_INS_AESD
 	ARM_INS_AESE      = C.ARM_INS_AESE
 	ARM_INS_AESIMC    = C.ARM_INS_AESIMC
 	ARM_INS_AESMC     = C.ARM_INS_AESMC
 	ARM_INS_AND       = C.ARM_INS_AND
+	ARM_INS_ASR       = C.ARM_INS_ASR
+	ARM_INS_B         = C.ARM_INS_B
 	ARM_INS_BFC       = C.ARM_INS_BFC
 	ARM_INS_BFI       = C.ARM_INS_BFI
 	ARM_INS_BIC       = C.ARM_INS_BIC
 	ARM_INS_BKPT      = C.ARM_INS_BKPT
 	ARM_INS_BL        = C.ARM_INS_BL
 	ARM_INS_BLX       = C.ARM_INS_BLX
+	ARM_INS_BLXNS     = C.ARM_INS_BLXNS
 	ARM_INS_BX        = C.ARM_INS_BX
 	ARM_INS_BXJ       = C.ARM_INS_BXJ
-	ARM_INS_B         = C.ARM_INS_B
+	ARM_INS_BXNS      = C.ARM_INS_BXNS
+	ARM_INS_CBNZ      = C.ARM_INS_CBNZ
+	ARM_INS_CBZ       = C.ARM_INS_CBZ
 	ARM_INS_CDP       = C.ARM_INS_CDP
 	ARM_INS_CDP2      = C.ARM_INS_CDP2
 	ARM_INS_CLREX     = C.ARM_INS_CLREX
@@ -374,21 +396,37 @@ const (
 	ARM_INS_CRC32CW   = C.ARM_INS_CRC32CW
 	ARM_INS_CRC32H    = C.ARM_INS_CRC32H
 	ARM_INS_CRC32W    = C.ARM_INS_CRC32W
+	ARM_INS_CSDB      = C.ARM_INS_CSDB
 	ARM_INS_DBG       = C.ARM_INS_DBG
+	ARM_INS_DCPS1     = C.ARM_INS_DCPS1
+	ARM_INS_DCPS2     = C.ARM_INS_DCPS2
+	ARM_INS_DCPS3     = C.ARM_INS_DCPS3
+	ARM_INS_DFB       = C.ARM_INS_DFB
 	ARM_INS_DMB       = C.ARM_INS_DMB
 	ARM_INS_DSB       = C.ARM_INS_DSB
 	ARM_INS_EOR       = C.ARM_INS_EOR
 	ARM_INS_ERET      = C.ARM_INS_ERET
-	ARM_INS_VMOV      = C.ARM_INS_VMOV
+	ARM_INS_ESB       = C.ARM_INS_ESB
+	ARM_INS_FADDD     = C.ARM_INS_FADDD
+	ARM_INS_FADDS     = C.ARM_INS_FADDS
+	ARM_INS_FCMPZD    = C.ARM_INS_FCMPZD
+	ARM_INS_FCMPZS    = C.ARM_INS_FCMPZS
+	ARM_INS_FCONSTD   = C.ARM_INS_FCONSTD
+	ARM_INS_FCONSTS   = C.ARM_INS_FCONSTS
 	ARM_INS_FLDMDBX   = C.ARM_INS_FLDMDBX
 	ARM_INS_FLDMIAX   = C.ARM_INS_FLDMIAX
-	ARM_INS_VMRS      = C.ARM_INS_VMRS
+	ARM_INS_FMDHR     = C.ARM_INS_FMDHR
+	ARM_INS_FMDLR     = C.ARM_INS_FMDLR
+	ARM_INS_FMSTAT    = C.ARM_INS_FMSTAT
 	ARM_INS_FSTMDBX   = C.ARM_INS_FSTMDBX
 	ARM_INS_FSTMIAX   = C.ARM_INS_FSTMIAX
+	ARM_INS_FSUBD     = C.ARM_INS_FSUBD
+	ARM_INS_FSUBS     = C.ARM_INS_FSUBS
 	ARM_INS_HINT      = C.ARM_INS_HINT
 	ARM_INS_HLT       = C.ARM_INS_HLT
 	ARM_INS_HVC       = C.ARM_INS_HVC
 	ARM_INS_ISB       = C.ARM_INS_ISB
+	ARM_INS_IT        = C.ARM_INS_IT
 	ARM_INS_LDA       = C.ARM_INS_LDA
 	ARM_INS_LDAB      = C.ARM_INS_LDAB
 	ARM_INS_LDAEX     = C.ARM_INS_LDAEX
@@ -396,16 +434,17 @@ const (
 	ARM_INS_LDAEXD    = C.ARM_INS_LDAEXD
 	ARM_INS_LDAEXH    = C.ARM_INS_LDAEXH
 	ARM_INS_LDAH      = C.ARM_INS_LDAH
-	ARM_INS_LDC2L     = C.ARM_INS_LDC2L
-	ARM_INS_LDC2      = C.ARM_INS_LDC2
-	ARM_INS_LDCL      = C.ARM_INS_LDCL
 	ARM_INS_LDC       = C.ARM_INS_LDC
+	ARM_INS_LDC2      = C.ARM_INS_LDC2
+	ARM_INS_LDC2L     = C.ARM_INS_LDC2L
+	ARM_INS_LDCL      = C.ARM_INS_LDCL
+	ARM_INS_LDM       = C.ARM_INS_LDM
 	ARM_INS_LDMDA     = C.ARM_INS_LDMDA
 	ARM_INS_LDMDB     = C.ARM_INS_LDMDB
-	ARM_INS_LDM       = C.ARM_INS_LDM
 	ARM_INS_LDMIB     = C.ARM_INS_LDMIB
-	ARM_INS_LDRBT     = C.ARM_INS_LDRBT
+	ARM_INS_LDR       = C.ARM_INS_LDR
 	ARM_INS_LDRB      = C.ARM_INS_LDRB
+	ARM_INS_LDRBT     = C.ARM_INS_LDRBT
 	ARM_INS_LDRD      = C.ARM_INS_LDRD
 	ARM_INS_LDREX     = C.ARM_INS_LDREX
 	ARM_INS_LDREXB    = C.ARM_INS_LDREXB
@@ -418,7 +457,8 @@ const (
 	ARM_INS_LDRSH     = C.ARM_INS_LDRSH
 	ARM_INS_LDRSHT    = C.ARM_INS_LDRSHT
 	ARM_INS_LDRT      = C.ARM_INS_LDRT
-	ARM_INS_LDR       = C.ARM_INS_LDR
+	ARM_INS_LSL       = C.ARM_INS_LSL
+	ARM_INS_LSR       = C.ARM_INS_LSR
 	ARM_INS_MCR       = C.ARM_INS_MCR
 	ARM_INS_MCR2      = C.ARM_INS_MCR2
 	ARM_INS_MCRR      = C.ARM_INS_MCRR
@@ -426,6 +466,7 @@ const (
 	ARM_INS_MLA       = C.ARM_INS_MLA
 	ARM_INS_MLS       = C.ARM_INS_MLS
 	ARM_INS_MOV       = C.ARM_INS_MOV
+	ARM_INS_MOVS      = C.ARM_INS_MOVS
 	ARM_INS_MOVT      = C.ARM_INS_MOVT
 	ARM_INS_MOVW      = C.ARM_INS_MOVW
 	ARM_INS_MRC       = C.ARM_INS_MRC
@@ -436,12 +477,17 @@ const (
 	ARM_INS_MSR       = C.ARM_INS_MSR
 	ARM_INS_MUL       = C.ARM_INS_MUL
 	ARM_INS_MVN       = C.ARM_INS_MVN
+	ARM_INS_NEG       = C.ARM_INS_NEG
+	ARM_INS_NOP       = C.ARM_INS_NOP
+	ARM_INS_ORN       = C.ARM_INS_ORN
 	ARM_INS_ORR       = C.ARM_INS_ORR
 	ARM_INS_PKHBT     = C.ARM_INS_PKHBT
 	ARM_INS_PKHTB     = C.ARM_INS_PKHTB
-	ARM_INS_PLDW      = C.ARM_INS_PLDW
 	ARM_INS_PLD       = C.ARM_INS_PLD
+	ARM_INS_PLDW      = C.ARM_INS_PLDW
 	ARM_INS_PLI       = C.ARM_INS_PLI
+	ARM_INS_POP       = C.ARM_INS_POP
+	ARM_INS_PUSH      = C.ARM_INS_PUSH
 	ARM_INS_QADD      = C.ARM_INS_QADD
 	ARM_INS_QADD16    = C.ARM_INS_QADD16
 	ARM_INS_QADD8     = C.ARM_INS_QADD8
@@ -460,6 +506,8 @@ const (
 	ARM_INS_RFEDB     = C.ARM_INS_RFEDB
 	ARM_INS_RFEIA     = C.ARM_INS_RFEIA
 	ARM_INS_RFEIB     = C.ARM_INS_RFEIB
+	ARM_INS_ROR       = C.ARM_INS_ROR
+	ARM_INS_RRX       = C.ARM_INS_RRX
 	ARM_INS_RSB       = C.ARM_INS_RSB
 	ARM_INS_RSC       = C.ARM_INS_RSC
 	ARM_INS_SADD16    = C.ARM_INS_SADD16
@@ -470,6 +518,10 @@ const (
 	ARM_INS_SDIV      = C.ARM_INS_SDIV
 	ARM_INS_SEL       = C.ARM_INS_SEL
 	ARM_INS_SETEND    = C.ARM_INS_SETEND
+	ARM_INS_SETPAN    = C.ARM_INS_SETPAN
+	ARM_INS_SEV       = C.ARM_INS_SEV
+	ARM_INS_SEVL      = C.ARM_INS_SEVL
+	ARM_INS_SG        = C.ARM_INS_SG
 	ARM_INS_SHA1C     = C.ARM_INS_SHA1C
 	ARM_INS_SHA1H     = C.ARM_INS_SHA1H
 	ARM_INS_SHA1M     = C.ARM_INS_SHA1M
@@ -532,10 +584,10 @@ const (
 	ARM_INS_SSAX      = C.ARM_INS_SSAX
 	ARM_INS_SSUB16    = C.ARM_INS_SSUB16
 	ARM_INS_SSUB8     = C.ARM_INS_SSUB8
-	ARM_INS_STC2L     = C.ARM_INS_STC2L
-	ARM_INS_STC2      = C.ARM_INS_STC2
-	ARM_INS_STCL      = C.ARM_INS_STCL
 	ARM_INS_STC       = C.ARM_INS_STC
+	ARM_INS_STC2      = C.ARM_INS_STC2
+	ARM_INS_STC2L     = C.ARM_INS_STC2L
+	ARM_INS_STCL      = C.ARM_INS_STCL
 	ARM_INS_STL       = C.ARM_INS_STL
 	ARM_INS_STLB      = C.ARM_INS_STLB
 	ARM_INS_STLEX     = C.ARM_INS_STLEX
@@ -543,12 +595,13 @@ const (
 	ARM_INS_STLEXD    = C.ARM_INS_STLEXD
 	ARM_INS_STLEXH    = C.ARM_INS_STLEXH
 	ARM_INS_STLH      = C.ARM_INS_STLH
+	ARM_INS_STM       = C.ARM_INS_STM
 	ARM_INS_STMDA     = C.ARM_INS_STMDA
 	ARM_INS_STMDB     = C.ARM_INS_STMDB
-	ARM_INS_STM       = C.ARM_INS_STM
 	ARM_INS_STMIB     = C.ARM_INS_STMIB
-	ARM_INS_STRBT     = C.ARM_INS_STRBT
+	ARM_INS_STR       = C.ARM_INS_STR
 	ARM_INS_STRB      = C.ARM_INS_STRB
+	ARM_INS_STRBT     = C.ARM_INS_STRBT
 	ARM_INS_STRD      = C.ARM_INS_STRD
 	ARM_INS_STREX     = C.ARM_INS_STREX
 	ARM_INS_STREXB    = C.ARM_INS_STREXB
@@ -557,8 +610,9 @@ const (
 	ARM_INS_STRH      = C.ARM_INS_STRH
 	ARM_INS_STRHT     = C.ARM_INS_STRHT
 	ARM_INS_STRT      = C.ARM_INS_STRT
-	ARM_INS_STR       = C.ARM_INS_STR
 	ARM_INS_SUB       = C.ARM_INS_SUB
+	ARM_INS_SUBS      = C.ARM_INS_SUBS
+	ARM_INS_SUBW      = C.ARM_INS_SUBW
 	ARM_INS_SVC       = C.ARM_INS_SVC
 	ARM_INS_SWP       = C.ARM_INS_SWP
 	ARM_INS_SWPB      = C.ARM_INS_SWPB
@@ -568,9 +622,16 @@ const (
 	ARM_INS_SXTB      = C.ARM_INS_SXTB
 	ARM_INS_SXTB16    = C.ARM_INS_SXTB16
 	ARM_INS_SXTH      = C.ARM_INS_SXTH
+	ARM_INS_TBB       = C.ARM_INS_TBB
+	ARM_INS_TBH       = C.ARM_INS_TBH
 	ARM_INS_TEQ       = C.ARM_INS_TEQ
 	ARM_INS_TRAP      = C.ARM_INS_TRAP
+	ARM_INS_TSB       = C.ARM_INS_TSB
 	ARM_INS_TST       = C.ARM_INS_TST
+	ARM_INS_TT        = C.ARM_INS_TT
+	ARM_INS_TTA       = C.ARM_INS_TTA
+	ARM_INS_TTAT      = C.ARM_INS_TTAT
+	ARM_INS_TTT       = C.ARM_INS_TTT
 	ARM_INS_UADD16    = C.ARM_INS_UADD16
 	ARM_INS_UADD8     = C.ARM_INS_UADD8
 	ARM_INS_UASX      = C.ARM_INS_UASX
@@ -605,13 +666,15 @@ const (
 	ARM_INS_UXTB      = C.ARM_INS_UXTB
 	ARM_INS_UXTB16    = C.ARM_INS_UXTB16
 	ARM_INS_UXTH      = C.ARM_INS_UXTH
-	ARM_INS_VABAL     = C.ARM_INS_VABAL
 	ARM_INS_VABA      = C.ARM_INS_VABA
-	ARM_INS_VABDL     = C.ARM_INS_VABDL
+	ARM_INS_VABAL     = C.ARM_INS_VABAL
 	ARM_INS_VABD      = C.ARM_INS_VABD
+	ARM_INS_VABDL     = C.ARM_INS_VABDL
 	ARM_INS_VABS      = C.ARM_INS_VABS
 	ARM_INS_VACGE     = C.ARM_INS_VACGE
 	ARM_INS_VACGT     = C.ARM_INS_VACGT
+	ARM_INS_VACLE     = C.ARM_INS_VACLE
+	ARM_INS_VACLT     = C.ARM_INS_VACLT
 	ARM_INS_VADD      = C.ARM_INS_VADD
 	ARM_INS_VADDHN    = C.ARM_INS_VADDHN
 	ARM_INS_VADDL     = C.ARM_INS_VADDL
@@ -621,6 +684,7 @@ const (
 	ARM_INS_VBIF      = C.ARM_INS_VBIF
 	ARM_INS_VBIT      = C.ARM_INS_VBIT
 	ARM_INS_VBSL      = C.ARM_INS_VBSL
+	ARM_INS_VCADD     = C.ARM_INS_VCADD
 	ARM_INS_VCEQ      = C.ARM_INS_VCEQ
 	ARM_INS_VCGE      = C.ARM_INS_VCGE
 	ARM_INS_VCGT      = C.ARM_INS_VCGT
@@ -628,15 +692,17 @@ const (
 	ARM_INS_VCLS      = C.ARM_INS_VCLS
 	ARM_INS_VCLT      = C.ARM_INS_VCLT
 	ARM_INS_VCLZ      = C.ARM_INS_VCLZ
+	ARM_INS_VCMLA     = C.ARM_INS_VCMLA
 	ARM_INS_VCMP      = C.ARM_INS_VCMP
 	ARM_INS_VCMPE     = C.ARM_INS_VCMPE
 	ARM_INS_VCNT      = C.ARM_INS_VCNT
+	ARM_INS_VCVT      = C.ARM_INS_VCVT
 	ARM_INS_VCVTA     = C.ARM_INS_VCVTA
 	ARM_INS_VCVTB     = C.ARM_INS_VCVTB
-	ARM_INS_VCVT      = C.ARM_INS_VCVT
 	ARM_INS_VCVTM     = C.ARM_INS_VCVTM
 	ARM_INS_VCVTN     = C.ARM_INS_VCVTN
 	ARM_INS_VCVTP     = C.ARM_INS_VCVTP
+	ARM_INS_VCVTR     = C.ARM_INS_VCVTR
 	ARM_INS_VCVTT     = C.ARM_INS_VCVTT
 	ARM_INS_VDIV      = C.ARM_INS_VDIV
 	ARM_INS_VDUP      = C.ARM_INS_VDUP
@@ -648,6 +714,8 @@ const (
 	ARM_INS_VFNMS     = C.ARM_INS_VFNMS
 	ARM_INS_VHADD     = C.ARM_INS_VHADD
 	ARM_INS_VHSUB     = C.ARM_INS_VHSUB
+	ARM_INS_VINS      = C.ARM_INS_VINS
+	ARM_INS_VJCVT     = C.ARM_INS_VJCVT
 	ARM_INS_VLD1      = C.ARM_INS_VLD1
 	ARM_INS_VLD2      = C.ARM_INS_VLD2
 	ARM_INS_VLD3      = C.ARM_INS_VLD3
@@ -655,16 +723,21 @@ const (
 	ARM_INS_VLDMDB    = C.ARM_INS_VLDMDB
 	ARM_INS_VLDMIA    = C.ARM_INS_VLDMIA
 	ARM_INS_VLDR      = C.ARM_INS_VLDR
-	ARM_INS_VMAXNM    = C.ARM_INS_VMAXNM
+	ARM_INS_VLLDM     = C.ARM_INS_VLLDM
+	ARM_INS_VLSTM     = C.ARM_INS_VLSTM
 	ARM_INS_VMAX      = C.ARM_INS_VMAX
-	ARM_INS_VMINNM    = C.ARM_INS_VMINNM
+	ARM_INS_VMAXNM    = C.ARM_INS_VMAXNM
 	ARM_INS_VMIN      = C.ARM_INS_VMIN
+	ARM_INS_VMINNM    = C.ARM_INS_VMINNM
 	ARM_INS_VMLA      = C.ARM_INS_VMLA
 	ARM_INS_VMLAL     = C.ARM_INS_VMLAL
 	ARM_INS_VMLS      = C.ARM_INS_VMLS
 	ARM_INS_VMLSL     = C.ARM_INS_VMLSL
+	ARM_INS_VMOV      = C.ARM_INS_VMOV
 	ARM_INS_VMOVL     = C.ARM_INS_VMOVL
 	ARM_INS_VMOVN     = C.ARM_INS_VMOVN
+	ARM_INS_VMOVX     = C.ARM_INS_VMOVX
+	ARM_INS_VMRS      = C.ARM_INS_VMRS
 	ARM_INS_VMSR      = C.ARM_INS_VMSR
 	ARM_INS_VMUL      = C.ARM_INS_VMUL
 	ARM_INS_VMULL     = C.ARM_INS_VMULL
@@ -676,19 +749,23 @@ const (
 	ARM_INS_VORN      = C.ARM_INS_VORN
 	ARM_INS_VORR      = C.ARM_INS_VORR
 	ARM_INS_VPADAL    = C.ARM_INS_VPADAL
-	ARM_INS_VPADDL    = C.ARM_INS_VPADDL
 	ARM_INS_VPADD     = C.ARM_INS_VPADD
+	ARM_INS_VPADDL    = C.ARM_INS_VPADDL
 	ARM_INS_VPMAX     = C.ARM_INS_VPMAX
 	ARM_INS_VPMIN     = C.ARM_INS_VPMIN
+	ARM_INS_VPOP      = C.ARM_INS_VPOP
+	ARM_INS_VPUSH     = C.ARM_INS_VPUSH
 	ARM_INS_VQABS     = C.ARM_INS_VQABS
 	ARM_INS_VQADD     = C.ARM_INS_VQADD
 	ARM_INS_VQDMLAL   = C.ARM_INS_VQDMLAL
 	ARM_INS_VQDMLSL   = C.ARM_INS_VQDMLSL
 	ARM_INS_VQDMULH   = C.ARM_INS_VQDMULH
 	ARM_INS_VQDMULL   = C.ARM_INS_VQDMULL
-	ARM_INS_VQMOVUN   = C.ARM_INS_VQMOVUN
 	ARM_INS_VQMOVN    = C.ARM_INS_VQMOVN
+	ARM_INS_VQMOVUN   = C.ARM_INS_VQMOVUN
 	ARM_INS_VQNEG     = C.ARM_INS_VQNEG
+	ARM_INS_VQRDMLAH  = C.ARM_INS_VQRDMLAH
+	ARM_INS_VQRDMLSH  = C.ARM_INS_VQRDMLSH
 	ARM_INS_VQRDMULH  = C.ARM_INS_VQRDMULH
 	ARM_INS_VQRSHL    = C.ARM_INS_VQRSHL
 	ARM_INS_VQRSHRN   = C.ARM_INS_VQRSHRN
@@ -713,20 +790,21 @@ const (
 	ARM_INS_VRINTX    = C.ARM_INS_VRINTX
 	ARM_INS_VRINTZ    = C.ARM_INS_VRINTZ
 	ARM_INS_VRSHL     = C.ARM_INS_VRSHL
-	ARM_INS_VRSHRN    = C.ARM_INS_VRSHRN
 	ARM_INS_VRSHR     = C.ARM_INS_VRSHR
+	ARM_INS_VRSHRN    = C.ARM_INS_VRSHRN
 	ARM_INS_VRSQRTE   = C.ARM_INS_VRSQRTE
 	ARM_INS_VRSQRTS   = C.ARM_INS_VRSQRTS
 	ARM_INS_VRSRA     = C.ARM_INS_VRSRA
 	ARM_INS_VRSUBHN   = C.ARM_INS_VRSUBHN
+	ARM_INS_VSDOT     = C.ARM_INS_VSDOT
 	ARM_INS_VSELEQ    = C.ARM_INS_VSELEQ
 	ARM_INS_VSELGE    = C.ARM_INS_VSELGE
 	ARM_INS_VSELGT    = C.ARM_INS_VSELGT
 	ARM_INS_VSELVS    = C.ARM_INS_VSELVS
-	ARM_INS_VSHLL     = C.ARM_INS_VSHLL
 	ARM_INS_VSHL      = C.ARM_INS_VSHL
-	ARM_INS_VSHRN     = C.ARM_INS_VSHRN
+	ARM_INS_VSHLL     = C.ARM_INS_VSHLL
 	ARM_INS_VSHR      = C.ARM_INS_VSHR
+	ARM_INS_VSHRN     = C.ARM_INS_VSHRN
 	ARM_INS_VSLI      = C.ARM_INS_VSLI
 	ARM_INS_VSQRT     = C.ARM_INS_VSQRT
 	ARM_INS_VSRA      = C.ARM_INS_VSRA
@@ -745,37 +823,14 @@ const (
 	ARM_INS_VSWP      = C.ARM_INS_VSWP
 	ARM_INS_VTBL      = C.ARM_INS_VTBL
 	ARM_INS_VTBX      = C.ARM_INS_VTBX
-	ARM_INS_VCVTR     = C.ARM_INS_VCVTR
 	ARM_INS_VTRN      = C.ARM_INS_VTRN
 	ARM_INS_VTST      = C.ARM_INS_VTST
+	ARM_INS_VUDOT     = C.ARM_INS_VUDOT
 	ARM_INS_VUZP      = C.ARM_INS_VUZP
 	ARM_INS_VZIP      = C.ARM_INS_VZIP
-	ARM_INS_ADDW      = C.ARM_INS_ADDW
-	ARM_INS_ASR       = C.ARM_INS_ASR
-	ARM_INS_DCPS1     = C.ARM_INS_DCPS1
-	ARM_INS_DCPS2     = C.ARM_INS_DCPS2
-	ARM_INS_DCPS3     = C.ARM_INS_DCPS3
-	ARM_INS_IT        = C.ARM_INS_IT
-	ARM_INS_LSL       = C.ARM_INS_LSL
-	ARM_INS_LSR       = C.ARM_INS_LSR
-	ARM_INS_ORN       = C.ARM_INS_ORN
-	ARM_INS_ROR       = C.ARM_INS_ROR
-	ARM_INS_RRX       = C.ARM_INS_RRX
-	ARM_INS_SUBW      = C.ARM_INS_SUBW
-	ARM_INS_TBB       = C.ARM_INS_TBB
-	ARM_INS_TBH       = C.ARM_INS_TBH
-	ARM_INS_CBNZ      = C.ARM_INS_CBNZ
-	ARM_INS_CBZ       = C.ARM_INS_CBZ
-	ARM_INS_POP       = C.ARM_INS_POP
-	ARM_INS_PUSH      = C.ARM_INS_PUSH
-	ARM_INS_NOP       = C.ARM_INS_NOP
-	ARM_INS_YIELD     = C.ARM_INS_YIELD
 	ARM_INS_WFE       = C.ARM_INS_WFE
 	ARM_INS_WFI       = C.ARM_INS_WFI
-	ARM_INS_SEV       = C.ARM_INS_SEV
-	ARM_INS_SEVL      = C.ARM_INS_SEVL
-	ARM_INS_VPUSH     = C.ARM_INS_VPUSH
-	ARM_INS_VPOP      = C.ARM_INS_VPOP
+	ARM_INS_YIELD     = C.ARM_INS_YIELD
 	ARM_INS_ENDING    = C.ARM_INS_ENDING
 )
 
