@@ -8,8 +8,8 @@ try reading the *_test.go files.
     (c) 2013 COSEINC. All Rights Reserved.
 
     THIS FILE WAS AUTO-GENERATED -- DO NOT EDIT!
-	Command: ./genconst ../capstone/bindings/python/capstone/
-	2021-10-14T08:50:33-04:00
+	Command: ./genconst ./capstone/bindings/python/capstone
+	2025-03-26T14:48:26+00:00
 
 */
 
@@ -54,6 +54,12 @@ const (
 const (
 	PPC_REG_INVALID = C.PPC_REG_INVALID
 	PPC_REG_CARRY   = C.PPC_REG_CARRY
+	PPC_REG_CTR     = C.PPC_REG_CTR
+	PPC_REG_LR      = C.PPC_REG_LR
+	PPC_REG_RM      = C.PPC_REG_RM
+	PPC_REG_VRSAVE  = C.PPC_REG_VRSAVE
+	PPC_REG_XER     = C.PPC_REG_XER
+	PPC_REG_ZERO    = C.PPC_REG_ZERO
 	PPC_REG_CR0     = C.PPC_REG_CR0
 	PPC_REG_CR1     = C.PPC_REG_CR1
 	PPC_REG_CR2     = C.PPC_REG_CR2
@@ -62,7 +68,7 @@ const (
 	PPC_REG_CR5     = C.PPC_REG_CR5
 	PPC_REG_CR6     = C.PPC_REG_CR6
 	PPC_REG_CR7     = C.PPC_REG_CR7
-	PPC_REG_CTR     = C.PPC_REG_CTR
+	PPC_REG_CTR8    = C.PPC_REG_CTR8
 	PPC_REG_F0      = C.PPC_REG_F0
 	PPC_REG_F1      = C.PPC_REG_F1
 	PPC_REG_F2      = C.PPC_REG_F2
@@ -95,7 +101,39 @@ const (
 	PPC_REG_F29     = C.PPC_REG_F29
 	PPC_REG_F30     = C.PPC_REG_F30
 	PPC_REG_F31     = C.PPC_REG_F31
-	PPC_REG_LR      = C.PPC_REG_LR
+	PPC_REG_LR8     = C.PPC_REG_LR8
+	PPC_REG_Q0      = C.PPC_REG_Q0
+	PPC_REG_Q1      = C.PPC_REG_Q1
+	PPC_REG_Q2      = C.PPC_REG_Q2
+	PPC_REG_Q3      = C.PPC_REG_Q3
+	PPC_REG_Q4      = C.PPC_REG_Q4
+	PPC_REG_Q5      = C.PPC_REG_Q5
+	PPC_REG_Q6      = C.PPC_REG_Q6
+	PPC_REG_Q7      = C.PPC_REG_Q7
+	PPC_REG_Q8      = C.PPC_REG_Q8
+	PPC_REG_Q9      = C.PPC_REG_Q9
+	PPC_REG_Q10     = C.PPC_REG_Q10
+	PPC_REG_Q11     = C.PPC_REG_Q11
+	PPC_REG_Q12     = C.PPC_REG_Q12
+	PPC_REG_Q13     = C.PPC_REG_Q13
+	PPC_REG_Q14     = C.PPC_REG_Q14
+	PPC_REG_Q15     = C.PPC_REG_Q15
+	PPC_REG_Q16     = C.PPC_REG_Q16
+	PPC_REG_Q17     = C.PPC_REG_Q17
+	PPC_REG_Q18     = C.PPC_REG_Q18
+	PPC_REG_Q19     = C.PPC_REG_Q19
+	PPC_REG_Q20     = C.PPC_REG_Q20
+	PPC_REG_Q21     = C.PPC_REG_Q21
+	PPC_REG_Q22     = C.PPC_REG_Q22
+	PPC_REG_Q23     = C.PPC_REG_Q23
+	PPC_REG_Q24     = C.PPC_REG_Q24
+	PPC_REG_Q25     = C.PPC_REG_Q25
+	PPC_REG_Q26     = C.PPC_REG_Q26
+	PPC_REG_Q27     = C.PPC_REG_Q27
+	PPC_REG_Q28     = C.PPC_REG_Q28
+	PPC_REG_Q29     = C.PPC_REG_Q29
+	PPC_REG_Q30     = C.PPC_REG_Q30
+	PPC_REG_Q31     = C.PPC_REG_Q31
 	PPC_REG_R0      = C.PPC_REG_R0
 	PPC_REG_R1      = C.PPC_REG_R1
 	PPC_REG_R2      = C.PPC_REG_R2
@@ -160,7 +198,6 @@ const (
 	PPC_REG_V29     = C.PPC_REG_V29
 	PPC_REG_V30     = C.PPC_REG_V30
 	PPC_REG_V31     = C.PPC_REG_V31
-	PPC_REG_VRSAVE  = C.PPC_REG_VRSAVE
 	PPC_REG_VS0     = C.PPC_REG_VS0
 	PPC_REG_VS1     = C.PPC_REG_VS1
 	PPC_REG_VS2     = C.PPC_REG_VS2
@@ -225,43 +262,38 @@ const (
 	PPC_REG_VS61    = C.PPC_REG_VS61
 	PPC_REG_VS62    = C.PPC_REG_VS62
 	PPC_REG_VS63    = C.PPC_REG_VS63
-	PPC_REG_Q0      = C.PPC_REG_Q0
-	PPC_REG_Q1      = C.PPC_REG_Q1
-	PPC_REG_Q2      = C.PPC_REG_Q2
-	PPC_REG_Q3      = C.PPC_REG_Q3
-	PPC_REG_Q4      = C.PPC_REG_Q4
-	PPC_REG_Q5      = C.PPC_REG_Q5
-	PPC_REG_Q6      = C.PPC_REG_Q6
-	PPC_REG_Q7      = C.PPC_REG_Q7
-	PPC_REG_Q8      = C.PPC_REG_Q8
-	PPC_REG_Q9      = C.PPC_REG_Q9
-	PPC_REG_Q10     = C.PPC_REG_Q10
-	PPC_REG_Q11     = C.PPC_REG_Q11
-	PPC_REG_Q12     = C.PPC_REG_Q12
-	PPC_REG_Q13     = C.PPC_REG_Q13
-	PPC_REG_Q14     = C.PPC_REG_Q14
-	PPC_REG_Q15     = C.PPC_REG_Q15
-	PPC_REG_Q16     = C.PPC_REG_Q16
-	PPC_REG_Q17     = C.PPC_REG_Q17
-	PPC_REG_Q18     = C.PPC_REG_Q18
-	PPC_REG_Q19     = C.PPC_REG_Q19
-	PPC_REG_Q20     = C.PPC_REG_Q20
-	PPC_REG_Q21     = C.PPC_REG_Q21
-	PPC_REG_Q22     = C.PPC_REG_Q22
-	PPC_REG_Q23     = C.PPC_REG_Q23
-	PPC_REG_Q24     = C.PPC_REG_Q24
-	PPC_REG_Q25     = C.PPC_REG_Q25
-	PPC_REG_Q26     = C.PPC_REG_Q26
-	PPC_REG_Q27     = C.PPC_REG_Q27
-	PPC_REG_Q28     = C.PPC_REG_Q28
-	PPC_REG_Q29     = C.PPC_REG_Q29
-	PPC_REG_Q30     = C.PPC_REG_Q30
-	PPC_REG_Q31     = C.PPC_REG_Q31
-	PPC_REG_RM      = C.PPC_REG_RM
-	PPC_REG_CTR8    = C.PPC_REG_CTR8
-	PPC_REG_LR8     = C.PPC_REG_LR8
+	PPC_REG_CR0EQ   = C.PPC_REG_CR0EQ
 	PPC_REG_CR1EQ   = C.PPC_REG_CR1EQ
-	PPC_REG_X2      = C.PPC_REG_X2
+	PPC_REG_CR2EQ   = C.PPC_REG_CR2EQ
+	PPC_REG_CR3EQ   = C.PPC_REG_CR3EQ
+	PPC_REG_CR4EQ   = C.PPC_REG_CR4EQ
+	PPC_REG_CR5EQ   = C.PPC_REG_CR5EQ
+	PPC_REG_CR6EQ   = C.PPC_REG_CR6EQ
+	PPC_REG_CR7EQ   = C.PPC_REG_CR7EQ
+	PPC_REG_CR0GT   = C.PPC_REG_CR0GT
+	PPC_REG_CR1GT   = C.PPC_REG_CR1GT
+	PPC_REG_CR2GT   = C.PPC_REG_CR2GT
+	PPC_REG_CR3GT   = C.PPC_REG_CR3GT
+	PPC_REG_CR4GT   = C.PPC_REG_CR4GT
+	PPC_REG_CR5GT   = C.PPC_REG_CR5GT
+	PPC_REG_CR6GT   = C.PPC_REG_CR6GT
+	PPC_REG_CR7GT   = C.PPC_REG_CR7GT
+	PPC_REG_CR0LT   = C.PPC_REG_CR0LT
+	PPC_REG_CR1LT   = C.PPC_REG_CR1LT
+	PPC_REG_CR2LT   = C.PPC_REG_CR2LT
+	PPC_REG_CR3LT   = C.PPC_REG_CR3LT
+	PPC_REG_CR4LT   = C.PPC_REG_CR4LT
+	PPC_REG_CR5LT   = C.PPC_REG_CR5LT
+	PPC_REG_CR6LT   = C.PPC_REG_CR6LT
+	PPC_REG_CR7LT   = C.PPC_REG_CR7LT
+	PPC_REG_CR0UN   = C.PPC_REG_CR0UN
+	PPC_REG_CR1UN   = C.PPC_REG_CR1UN
+	PPC_REG_CR2UN   = C.PPC_REG_CR2UN
+	PPC_REG_CR3UN   = C.PPC_REG_CR3UN
+	PPC_REG_CR4UN   = C.PPC_REG_CR4UN
+	PPC_REG_CR5UN   = C.PPC_REG_CR5UN
+	PPC_REG_CR6UN   = C.PPC_REG_CR6UN
+	PPC_REG_CR7UN   = C.PPC_REG_CR7UN
 	PPC_REG_ENDING  = C.PPC_REG_ENDING
 )
 
@@ -274,71 +306,265 @@ const (
 	PPC_INS_ADDIC         = C.PPC_INS_ADDIC
 	PPC_INS_ADDIS         = C.PPC_INS_ADDIS
 	PPC_INS_ADDME         = C.PPC_INS_ADDME
+	PPC_INS_ADDPCIS       = C.PPC_INS_ADDPCIS
 	PPC_INS_ADDZE         = C.PPC_INS_ADDZE
 	PPC_INS_AND           = C.PPC_INS_AND
 	PPC_INS_ANDC          = C.PPC_INS_ANDC
-	PPC_INS_ANDIS         = C.PPC_INS_ANDIS
 	PPC_INS_ANDI          = C.PPC_INS_ANDI
+	PPC_INS_ANDIS         = C.PPC_INS_ANDIS
 	PPC_INS_ATTN          = C.PPC_INS_ATTN
 	PPC_INS_B             = C.PPC_INS_B
 	PPC_INS_BA            = C.PPC_INS_BA
 	PPC_INS_BC            = C.PPC_INS_BC
+	PPC_INS_BCA           = C.PPC_INS_BCA
 	PPC_INS_BCCTR         = C.PPC_INS_BCCTR
 	PPC_INS_BCCTRL        = C.PPC_INS_BCCTRL
+	PPC_INS_BCDCFN        = C.PPC_INS_BCDCFN
+	PPC_INS_BCDCFSQ       = C.PPC_INS_BCDCFSQ
+	PPC_INS_BCDCFZ        = C.PPC_INS_BCDCFZ
+	PPC_INS_BCDCPSGN      = C.PPC_INS_BCDCPSGN
+	PPC_INS_BCDCTN        = C.PPC_INS_BCDCTN
+	PPC_INS_BCDCTSQ       = C.PPC_INS_BCDCTSQ
+	PPC_INS_BCDCTZ        = C.PPC_INS_BCDCTZ
+	PPC_INS_BCDS          = C.PPC_INS_BCDS
+	PPC_INS_BCDSETSGN     = C.PPC_INS_BCDSETSGN
+	PPC_INS_BCDSR         = C.PPC_INS_BCDSR
+	PPC_INS_BCDTRUNC      = C.PPC_INS_BCDTRUNC
+	PPC_INS_BCDUS         = C.PPC_INS_BCDUS
+	PPC_INS_BCDUTRUNC     = C.PPC_INS_BCDUTRUNC
 	PPC_INS_BCL           = C.PPC_INS_BCL
+	PPC_INS_BCLA          = C.PPC_INS_BCLA
 	PPC_INS_BCLR          = C.PPC_INS_BCLR
 	PPC_INS_BCLRL         = C.PPC_INS_BCLRL
 	PPC_INS_BCTR          = C.PPC_INS_BCTR
 	PPC_INS_BCTRL         = C.PPC_INS_BCTRL
-	PPC_INS_BCT           = C.PPC_INS_BCT
 	PPC_INS_BDNZ          = C.PPC_INS_BDNZ
 	PPC_INS_BDNZA         = C.PPC_INS_BDNZA
+	PPC_INS_BDNZF         = C.PPC_INS_BDNZF
+	PPC_INS_BDNZFA        = C.PPC_INS_BDNZFA
+	PPC_INS_BDNZFL        = C.PPC_INS_BDNZFL
+	PPC_INS_BDNZFLA       = C.PPC_INS_BDNZFLA
+	PPC_INS_BDNZFLR       = C.PPC_INS_BDNZFLR
+	PPC_INS_BDNZFLRL      = C.PPC_INS_BDNZFLRL
 	PPC_INS_BDNZL         = C.PPC_INS_BDNZL
 	PPC_INS_BDNZLA        = C.PPC_INS_BDNZLA
 	PPC_INS_BDNZLR        = C.PPC_INS_BDNZLR
 	PPC_INS_BDNZLRL       = C.PPC_INS_BDNZLRL
+	PPC_INS_BDNZT         = C.PPC_INS_BDNZT
+	PPC_INS_BDNZTA        = C.PPC_INS_BDNZTA
+	PPC_INS_BDNZTL        = C.PPC_INS_BDNZTL
+	PPC_INS_BDNZTLA       = C.PPC_INS_BDNZTLA
+	PPC_INS_BDNZTLR       = C.PPC_INS_BDNZTLR
+	PPC_INS_BDNZTLRL      = C.PPC_INS_BDNZTLRL
 	PPC_INS_BDZ           = C.PPC_INS_BDZ
 	PPC_INS_BDZA          = C.PPC_INS_BDZA
+	PPC_INS_BDZF          = C.PPC_INS_BDZF
+	PPC_INS_BDZFA         = C.PPC_INS_BDZFA
+	PPC_INS_BDZFL         = C.PPC_INS_BDZFL
+	PPC_INS_BDZFLA        = C.PPC_INS_BDZFLA
+	PPC_INS_BDZFLR        = C.PPC_INS_BDZFLR
+	PPC_INS_BDZFLRL       = C.PPC_INS_BDZFLRL
 	PPC_INS_BDZL          = C.PPC_INS_BDZL
 	PPC_INS_BDZLA         = C.PPC_INS_BDZLA
 	PPC_INS_BDZLR         = C.PPC_INS_BDZLR
 	PPC_INS_BDZLRL        = C.PPC_INS_BDZLRL
+	PPC_INS_BDZT          = C.PPC_INS_BDZT
+	PPC_INS_BDZTA         = C.PPC_INS_BDZTA
+	PPC_INS_BDZTL         = C.PPC_INS_BDZTL
+	PPC_INS_BDZTLA        = C.PPC_INS_BDZTLA
+	PPC_INS_BDZTLR        = C.PPC_INS_BDZTLR
+	PPC_INS_BDZTLRL       = C.PPC_INS_BDZTLRL
+	PPC_INS_BEQ           = C.PPC_INS_BEQ
+	PPC_INS_BEQA          = C.PPC_INS_BEQA
+	PPC_INS_BEQCTR        = C.PPC_INS_BEQCTR
+	PPC_INS_BEQCTRL       = C.PPC_INS_BEQCTRL
+	PPC_INS_BEQL          = C.PPC_INS_BEQL
+	PPC_INS_BEQLA         = C.PPC_INS_BEQLA
+	PPC_INS_BEQLR         = C.PPC_INS_BEQLR
+	PPC_INS_BEQLRL        = C.PPC_INS_BEQLRL
+	PPC_INS_BF            = C.PPC_INS_BF
+	PPC_INS_BFA           = C.PPC_INS_BFA
+	PPC_INS_BFCTR         = C.PPC_INS_BFCTR
+	PPC_INS_BFCTRL        = C.PPC_INS_BFCTRL
+	PPC_INS_BFL           = C.PPC_INS_BFL
+	PPC_INS_BFLA          = C.PPC_INS_BFLA
+	PPC_INS_BFLR          = C.PPC_INS_BFLR
+	PPC_INS_BFLRL         = C.PPC_INS_BFLRL
+	PPC_INS_BGE           = C.PPC_INS_BGE
+	PPC_INS_BGEA          = C.PPC_INS_BGEA
+	PPC_INS_BGECTR        = C.PPC_INS_BGECTR
+	PPC_INS_BGECTRL       = C.PPC_INS_BGECTRL
+	PPC_INS_BGEL          = C.PPC_INS_BGEL
+	PPC_INS_BGELA         = C.PPC_INS_BGELA
+	PPC_INS_BGELR         = C.PPC_INS_BGELR
+	PPC_INS_BGELRL        = C.PPC_INS_BGELRL
+	PPC_INS_BGT           = C.PPC_INS_BGT
+	PPC_INS_BGTA          = C.PPC_INS_BGTA
+	PPC_INS_BGTCTR        = C.PPC_INS_BGTCTR
+	PPC_INS_BGTCTRL       = C.PPC_INS_BGTCTRL
+	PPC_INS_BGTL          = C.PPC_INS_BGTL
+	PPC_INS_BGTLA         = C.PPC_INS_BGTLA
+	PPC_INS_BGTLR         = C.PPC_INS_BGTLR
+	PPC_INS_BGTLRL        = C.PPC_INS_BGTLRL
 	PPC_INS_BL            = C.PPC_INS_BL
 	PPC_INS_BLA           = C.PPC_INS_BLA
+	PPC_INS_BLE           = C.PPC_INS_BLE
+	PPC_INS_BLEA          = C.PPC_INS_BLEA
+	PPC_INS_BLECTR        = C.PPC_INS_BLECTR
+	PPC_INS_BLECTRL       = C.PPC_INS_BLECTRL
+	PPC_INS_BLEL          = C.PPC_INS_BLEL
+	PPC_INS_BLELA         = C.PPC_INS_BLELA
+	PPC_INS_BLELR         = C.PPC_INS_BLELR
+	PPC_INS_BLELRL        = C.PPC_INS_BLELRL
 	PPC_INS_BLR           = C.PPC_INS_BLR
 	PPC_INS_BLRL          = C.PPC_INS_BLRL
+	PPC_INS_BLT           = C.PPC_INS_BLT
+	PPC_INS_BLTA          = C.PPC_INS_BLTA
+	PPC_INS_BLTCTR        = C.PPC_INS_BLTCTR
+	PPC_INS_BLTCTRL       = C.PPC_INS_BLTCTRL
+	PPC_INS_BLTL          = C.PPC_INS_BLTL
+	PPC_INS_BLTLA         = C.PPC_INS_BLTLA
+	PPC_INS_BLTLR         = C.PPC_INS_BLTLR
+	PPC_INS_BLTLRL        = C.PPC_INS_BLTLRL
+	PPC_INS_BNE           = C.PPC_INS_BNE
+	PPC_INS_BNEA          = C.PPC_INS_BNEA
+	PPC_INS_BNECTR        = C.PPC_INS_BNECTR
+	PPC_INS_BNECTRL       = C.PPC_INS_BNECTRL
+	PPC_INS_BNEL          = C.PPC_INS_BNEL
+	PPC_INS_BNELA         = C.PPC_INS_BNELA
+	PPC_INS_BNELR         = C.PPC_INS_BNELR
+	PPC_INS_BNELRL        = C.PPC_INS_BNELRL
+	PPC_INS_BNG           = C.PPC_INS_BNG
+	PPC_INS_BNGA          = C.PPC_INS_BNGA
+	PPC_INS_BNGCTR        = C.PPC_INS_BNGCTR
+	PPC_INS_BNGCTRL       = C.PPC_INS_BNGCTRL
+	PPC_INS_BNGL          = C.PPC_INS_BNGL
+	PPC_INS_BNGLA         = C.PPC_INS_BNGLA
+	PPC_INS_BNGLR         = C.PPC_INS_BNGLR
+	PPC_INS_BNGLRL        = C.PPC_INS_BNGLRL
+	PPC_INS_BNL           = C.PPC_INS_BNL
+	PPC_INS_BNLA          = C.PPC_INS_BNLA
+	PPC_INS_BNLCTR        = C.PPC_INS_BNLCTR
+	PPC_INS_BNLCTRL       = C.PPC_INS_BNLCTRL
+	PPC_INS_BNLL          = C.PPC_INS_BNLL
+	PPC_INS_BNLLA         = C.PPC_INS_BNLLA
+	PPC_INS_BNLLR         = C.PPC_INS_BNLLR
+	PPC_INS_BNLLRL        = C.PPC_INS_BNLLRL
+	PPC_INS_BNS           = C.PPC_INS_BNS
+	PPC_INS_BNSA          = C.PPC_INS_BNSA
+	PPC_INS_BNSCTR        = C.PPC_INS_BNSCTR
+	PPC_INS_BNSCTRL       = C.PPC_INS_BNSCTRL
+	PPC_INS_BNSL          = C.PPC_INS_BNSL
+	PPC_INS_BNSLA         = C.PPC_INS_BNSLA
+	PPC_INS_BNSLR         = C.PPC_INS_BNSLR
+	PPC_INS_BNSLRL        = C.PPC_INS_BNSLRL
+	PPC_INS_BNU           = C.PPC_INS_BNU
+	PPC_INS_BNUA          = C.PPC_INS_BNUA
+	PPC_INS_BNUCTR        = C.PPC_INS_BNUCTR
+	PPC_INS_BNUCTRL       = C.PPC_INS_BNUCTRL
+	PPC_INS_BNUL          = C.PPC_INS_BNUL
+	PPC_INS_BNULA         = C.PPC_INS_BNULA
+	PPC_INS_BNULR         = C.PPC_INS_BNULR
+	PPC_INS_BNULRL        = C.PPC_INS_BNULRL
+	PPC_INS_BPERMD        = C.PPC_INS_BPERMD
 	PPC_INS_BRINC         = C.PPC_INS_BRINC
+	PPC_INS_BSO           = C.PPC_INS_BSO
+	PPC_INS_BSOA          = C.PPC_INS_BSOA
+	PPC_INS_BSOCTR        = C.PPC_INS_BSOCTR
+	PPC_INS_BSOCTRL       = C.PPC_INS_BSOCTRL
+	PPC_INS_BSOL          = C.PPC_INS_BSOL
+	PPC_INS_BSOLA         = C.PPC_INS_BSOLA
+	PPC_INS_BSOLR         = C.PPC_INS_BSOLR
+	PPC_INS_BSOLRL        = C.PPC_INS_BSOLRL
+	PPC_INS_BT            = C.PPC_INS_BT
+	PPC_INS_BTA           = C.PPC_INS_BTA
+	PPC_INS_BTCTR         = C.PPC_INS_BTCTR
+	PPC_INS_BTCTRL        = C.PPC_INS_BTCTRL
+	PPC_INS_BTL           = C.PPC_INS_BTL
+	PPC_INS_BTLA          = C.PPC_INS_BTLA
+	PPC_INS_BTLR          = C.PPC_INS_BTLR
+	PPC_INS_BTLRL         = C.PPC_INS_BTLRL
+	PPC_INS_BUN           = C.PPC_INS_BUN
+	PPC_INS_BUNA          = C.PPC_INS_BUNA
+	PPC_INS_BUNCTR        = C.PPC_INS_BUNCTR
+	PPC_INS_BUNCTRL       = C.PPC_INS_BUNCTRL
+	PPC_INS_BUNL          = C.PPC_INS_BUNL
+	PPC_INS_BUNLA         = C.PPC_INS_BUNLA
+	PPC_INS_BUNLR         = C.PPC_INS_BUNLR
+	PPC_INS_BUNLRL        = C.PPC_INS_BUNLRL
+	PPC_INS_CLRBHRB       = C.PPC_INS_CLRBHRB
+	PPC_INS_CLRLDI        = C.PPC_INS_CLRLDI
+	PPC_INS_CLRLSLDI      = C.PPC_INS_CLRLSLDI
+	PPC_INS_CLRLSLWI      = C.PPC_INS_CLRLSLWI
+	PPC_INS_CLRLWI        = C.PPC_INS_CLRLWI
+	PPC_INS_CLRRDI        = C.PPC_INS_CLRRDI
+	PPC_INS_CLRRWI        = C.PPC_INS_CLRRWI
+	PPC_INS_CMP           = C.PPC_INS_CMP
 	PPC_INS_CMPB          = C.PPC_INS_CMPB
 	PPC_INS_CMPD          = C.PPC_INS_CMPD
 	PPC_INS_CMPDI         = C.PPC_INS_CMPDI
+	PPC_INS_CMPEQB        = C.PPC_INS_CMPEQB
+	PPC_INS_CMPI          = C.PPC_INS_CMPI
+	PPC_INS_CMPL          = C.PPC_INS_CMPL
 	PPC_INS_CMPLD         = C.PPC_INS_CMPLD
 	PPC_INS_CMPLDI        = C.PPC_INS_CMPLDI
+	PPC_INS_CMPLI         = C.PPC_INS_CMPLI
 	PPC_INS_CMPLW         = C.PPC_INS_CMPLW
 	PPC_INS_CMPLWI        = C.PPC_INS_CMPLWI
+	PPC_INS_CMPRB         = C.PPC_INS_CMPRB
 	PPC_INS_CMPW          = C.PPC_INS_CMPW
 	PPC_INS_CMPWI         = C.PPC_INS_CMPWI
 	PPC_INS_CNTLZD        = C.PPC_INS_CNTLZD
 	PPC_INS_CNTLZW        = C.PPC_INS_CNTLZW
-	PPC_INS_CREQV         = C.PPC_INS_CREQV
-	PPC_INS_CRXOR         = C.PPC_INS_CRXOR
+	PPC_INS_CNTTZD        = C.PPC_INS_CNTTZD
+	PPC_INS_CNTTZW        = C.PPC_INS_CNTTZW
+	PPC_INS_COPY          = C.PPC_INS_COPY
+	PPC_INS_COPY_FIRST    = C.PPC_INS_COPY_FIRST
+	PPC_INS_CP_ABORT      = C.PPC_INS_CP_ABORT
 	PPC_INS_CRAND         = C.PPC_INS_CRAND
 	PPC_INS_CRANDC        = C.PPC_INS_CRANDC
+	PPC_INS_CRCLR         = C.PPC_INS_CRCLR
+	PPC_INS_CREQV         = C.PPC_INS_CREQV
+	PPC_INS_CRMOVE        = C.PPC_INS_CRMOVE
 	PPC_INS_CRNAND        = C.PPC_INS_CRNAND
 	PPC_INS_CRNOR         = C.PPC_INS_CRNOR
+	PPC_INS_CRNOT         = C.PPC_INS_CRNOT
 	PPC_INS_CROR          = C.PPC_INS_CROR
 	PPC_INS_CRORC         = C.PPC_INS_CRORC
+	PPC_INS_CRSET         = C.PPC_INS_CRSET
+	PPC_INS_CRXOR         = C.PPC_INS_CRXOR
+	PPC_INS_DARN          = C.PPC_INS_DARN
 	PPC_INS_DCBA          = C.PPC_INS_DCBA
 	PPC_INS_DCBF          = C.PPC_INS_DCBF
+	PPC_INS_DCBFEP        = C.PPC_INS_DCBFEP
+	PPC_INS_DCBFL         = C.PPC_INS_DCBFL
+	PPC_INS_DCBFLP        = C.PPC_INS_DCBFLP
 	PPC_INS_DCBI          = C.PPC_INS_DCBI
 	PPC_INS_DCBST         = C.PPC_INS_DCBST
+	PPC_INS_DCBSTEP       = C.PPC_INS_DCBSTEP
 	PPC_INS_DCBT          = C.PPC_INS_DCBT
+	PPC_INS_DCBTCT        = C.PPC_INS_DCBTCT
+	PPC_INS_DCBTDS        = C.PPC_INS_DCBTDS
+	PPC_INS_DCBTEP        = C.PPC_INS_DCBTEP
 	PPC_INS_DCBTST        = C.PPC_INS_DCBTST
+	PPC_INS_DCBTSTCT      = C.PPC_INS_DCBTSTCT
+	PPC_INS_DCBTSTDS      = C.PPC_INS_DCBTSTDS
+	PPC_INS_DCBTSTEP      = C.PPC_INS_DCBTSTEP
+	PPC_INS_DCBTSTT       = C.PPC_INS_DCBTSTT
+	PPC_INS_DCBTT         = C.PPC_INS_DCBTT
 	PPC_INS_DCBZ          = C.PPC_INS_DCBZ
+	PPC_INS_DCBZEP        = C.PPC_INS_DCBZEP
 	PPC_INS_DCBZL         = C.PPC_INS_DCBZL
+	PPC_INS_DCBZLEP       = C.PPC_INS_DCBZLEP
 	PPC_INS_DCCCI         = C.PPC_INS_DCCCI
+	PPC_INS_DCI           = C.PPC_INS_DCI
 	PPC_INS_DIVD          = C.PPC_INS_DIVD
+	PPC_INS_DIVDE         = C.PPC_INS_DIVDE
+	PPC_INS_DIVDEU        = C.PPC_INS_DIVDEU
 	PPC_INS_DIVDU         = C.PPC_INS_DIVDU
 	PPC_INS_DIVW          = C.PPC_INS_DIVW
+	PPC_INS_DIVWE         = C.PPC_INS_DIVWE
+	PPC_INS_DIVWEU        = C.PPC_INS_DIVWEU
 	PPC_INS_DIVWU         = C.PPC_INS_DIVWU
 	PPC_INS_DSS           = C.PPC_INS_DSS
 	PPC_INS_DSSALL        = C.PPC_INS_DSSALL
@@ -346,6 +572,59 @@ const (
 	PPC_INS_DSTST         = C.PPC_INS_DSTST
 	PPC_INS_DSTSTT        = C.PPC_INS_DSTSTT
 	PPC_INS_DSTT          = C.PPC_INS_DSTT
+	PPC_INS_EFDABS        = C.PPC_INS_EFDABS
+	PPC_INS_EFDADD        = C.PPC_INS_EFDADD
+	PPC_INS_EFDCFS        = C.PPC_INS_EFDCFS
+	PPC_INS_EFDCFSF       = C.PPC_INS_EFDCFSF
+	PPC_INS_EFDCFSI       = C.PPC_INS_EFDCFSI
+	PPC_INS_EFDCFSID      = C.PPC_INS_EFDCFSID
+	PPC_INS_EFDCFUF       = C.PPC_INS_EFDCFUF
+	PPC_INS_EFDCFUI       = C.PPC_INS_EFDCFUI
+	PPC_INS_EFDCFUID      = C.PPC_INS_EFDCFUID
+	PPC_INS_EFDCMPEQ      = C.PPC_INS_EFDCMPEQ
+	PPC_INS_EFDCMPGT      = C.PPC_INS_EFDCMPGT
+	PPC_INS_EFDCMPLT      = C.PPC_INS_EFDCMPLT
+	PPC_INS_EFDCTSF       = C.PPC_INS_EFDCTSF
+	PPC_INS_EFDCTSI       = C.PPC_INS_EFDCTSI
+	PPC_INS_EFDCTSIDZ     = C.PPC_INS_EFDCTSIDZ
+	PPC_INS_EFDCTSIZ      = C.PPC_INS_EFDCTSIZ
+	PPC_INS_EFDCTUF       = C.PPC_INS_EFDCTUF
+	PPC_INS_EFDCTUI       = C.PPC_INS_EFDCTUI
+	PPC_INS_EFDCTUIDZ     = C.PPC_INS_EFDCTUIDZ
+	PPC_INS_EFDCTUIZ      = C.PPC_INS_EFDCTUIZ
+	PPC_INS_EFDDIV        = C.PPC_INS_EFDDIV
+	PPC_INS_EFDMUL        = C.PPC_INS_EFDMUL
+	PPC_INS_EFDNABS       = C.PPC_INS_EFDNABS
+	PPC_INS_EFDNEG        = C.PPC_INS_EFDNEG
+	PPC_INS_EFDSUB        = C.PPC_INS_EFDSUB
+	PPC_INS_EFDTSTEQ      = C.PPC_INS_EFDTSTEQ
+	PPC_INS_EFDTSTGT      = C.PPC_INS_EFDTSTGT
+	PPC_INS_EFDTSTLT      = C.PPC_INS_EFDTSTLT
+	PPC_INS_EFSABS        = C.PPC_INS_EFSABS
+	PPC_INS_EFSADD        = C.PPC_INS_EFSADD
+	PPC_INS_EFSCFD        = C.PPC_INS_EFSCFD
+	PPC_INS_EFSCFSF       = C.PPC_INS_EFSCFSF
+	PPC_INS_EFSCFSI       = C.PPC_INS_EFSCFSI
+	PPC_INS_EFSCFUF       = C.PPC_INS_EFSCFUF
+	PPC_INS_EFSCFUI       = C.PPC_INS_EFSCFUI
+	PPC_INS_EFSCMPEQ      = C.PPC_INS_EFSCMPEQ
+	PPC_INS_EFSCMPGT      = C.PPC_INS_EFSCMPGT
+	PPC_INS_EFSCMPLT      = C.PPC_INS_EFSCMPLT
+	PPC_INS_EFSCTSF       = C.PPC_INS_EFSCTSF
+	PPC_INS_EFSCTSI       = C.PPC_INS_EFSCTSI
+	PPC_INS_EFSCTSIZ      = C.PPC_INS_EFSCTSIZ
+	PPC_INS_EFSCTUF       = C.PPC_INS_EFSCTUF
+	PPC_INS_EFSCTUI       = C.PPC_INS_EFSCTUI
+	PPC_INS_EFSCTUIZ      = C.PPC_INS_EFSCTUIZ
+	PPC_INS_EFSDIV        = C.PPC_INS_EFSDIV
+	PPC_INS_EFSMUL        = C.PPC_INS_EFSMUL
+	PPC_INS_EFSNABS       = C.PPC_INS_EFSNABS
+	PPC_INS_EFSNEG        = C.PPC_INS_EFSNEG
+	PPC_INS_EFSSUB        = C.PPC_INS_EFSSUB
+	PPC_INS_EFSTSTEQ      = C.PPC_INS_EFSTSTEQ
+	PPC_INS_EFSTSTGT      = C.PPC_INS_EFSTSTGT
+	PPC_INS_EFSTSTLT      = C.PPC_INS_EFSTSTLT
+	PPC_INS_EIEIO         = C.PPC_INS_EIEIO
 	PPC_INS_EQV           = C.PPC_INS_EQV
 	PPC_INS_EVABS         = C.PPC_INS_EVABS
 	PPC_INS_EVADDIW       = C.PPC_INS_EVADDIW
@@ -368,6 +647,27 @@ const (
 	PPC_INS_EVEQV         = C.PPC_INS_EVEQV
 	PPC_INS_EVEXTSB       = C.PPC_INS_EVEXTSB
 	PPC_INS_EVEXTSH       = C.PPC_INS_EVEXTSH
+	PPC_INS_EVFSABS       = C.PPC_INS_EVFSABS
+	PPC_INS_EVFSADD       = C.PPC_INS_EVFSADD
+	PPC_INS_EVFSCFSF      = C.PPC_INS_EVFSCFSF
+	PPC_INS_EVFSCFSI      = C.PPC_INS_EVFSCFSI
+	PPC_INS_EVFSCFUF      = C.PPC_INS_EVFSCFUF
+	PPC_INS_EVFSCFUI      = C.PPC_INS_EVFSCFUI
+	PPC_INS_EVFSCMPEQ     = C.PPC_INS_EVFSCMPEQ
+	PPC_INS_EVFSCMPGT     = C.PPC_INS_EVFSCMPGT
+	PPC_INS_EVFSCMPLT     = C.PPC_INS_EVFSCMPLT
+	PPC_INS_EVFSCTSF      = C.PPC_INS_EVFSCTSF
+	PPC_INS_EVFSCTSI      = C.PPC_INS_EVFSCTSI
+	PPC_INS_EVFSCTSIZ     = C.PPC_INS_EVFSCTSIZ
+	PPC_INS_EVFSCTUI      = C.PPC_INS_EVFSCTUI
+	PPC_INS_EVFSDIV       = C.PPC_INS_EVFSDIV
+	PPC_INS_EVFSMUL       = C.PPC_INS_EVFSMUL
+	PPC_INS_EVFSNABS      = C.PPC_INS_EVFSNABS
+	PPC_INS_EVFSNEG       = C.PPC_INS_EVFSNEG
+	PPC_INS_EVFSSUB       = C.PPC_INS_EVFSSUB
+	PPC_INS_EVFSTSTEQ     = C.PPC_INS_EVFSTSTEQ
+	PPC_INS_EVFSTSTGT     = C.PPC_INS_EVFSTSTGT
+	PPC_INS_EVFSTSTLT     = C.PPC_INS_EVFSTSTLT
 	PPC_INS_EVLDD         = C.PPC_INS_EVLDD
 	PPC_INS_EVLDDX        = C.PPC_INS_EVLDDX
 	PPC_INS_EVLDH         = C.PPC_INS_EVLDH
@@ -489,6 +789,7 @@ const (
 	PPC_INS_EVRLW         = C.PPC_INS_EVRLW
 	PPC_INS_EVRLWI        = C.PPC_INS_EVRLWI
 	PPC_INS_EVRNDW        = C.PPC_INS_EVRNDW
+	PPC_INS_EVSEL         = C.PPC_INS_EVSEL
 	PPC_INS_EVSLW         = C.PPC_INS_EVSLW
 	PPC_INS_EVSLWI        = C.PPC_INS_EVSLWI
 	PPC_INS_EVSPLATFI     = C.PPC_INS_EVSPLATFI
@@ -518,10 +819,14 @@ const (
 	PPC_INS_EVSUBFW       = C.PPC_INS_EVSUBFW
 	PPC_INS_EVSUBIFW      = C.PPC_INS_EVSUBIFW
 	PPC_INS_EVXOR         = C.PPC_INS_EVXOR
+	PPC_INS_EXTLDI        = C.PPC_INS_EXTLDI
+	PPC_INS_EXTLWI        = C.PPC_INS_EXTLWI
+	PPC_INS_EXTRDI        = C.PPC_INS_EXTRDI
+	PPC_INS_EXTRWI        = C.PPC_INS_EXTRWI
 	PPC_INS_EXTSB         = C.PPC_INS_EXTSB
 	PPC_INS_EXTSH         = C.PPC_INS_EXTSH
 	PPC_INS_EXTSW         = C.PPC_INS_EXTSW
-	PPC_INS_EIEIO         = C.PPC_INS_EIEIO
+	PPC_INS_EXTSWSLI      = C.PPC_INS_EXTSWSLI
 	PPC_INS_FABS          = C.PPC_INS_FABS
 	PPC_INS_FADD          = C.PPC_INS_FADD
 	PPC_INS_FADDS         = C.PPC_INS_FADDS
@@ -532,9 +837,11 @@ const (
 	PPC_INS_FCMPU         = C.PPC_INS_FCMPU
 	PPC_INS_FCPSGN        = C.PPC_INS_FCPSGN
 	PPC_INS_FCTID         = C.PPC_INS_FCTID
+	PPC_INS_FCTIDU        = C.PPC_INS_FCTIDU
 	PPC_INS_FCTIDUZ       = C.PPC_INS_FCTIDUZ
 	PPC_INS_FCTIDZ        = C.PPC_INS_FCTIDZ
 	PPC_INS_FCTIW         = C.PPC_INS_FCTIW
+	PPC_INS_FCTIWU        = C.PPC_INS_FCTIWU
 	PPC_INS_FCTIWUZ       = C.PPC_INS_FCTIWUZ
 	PPC_INS_FCTIWZ        = C.PPC_INS_FCTIWZ
 	PPC_INS_FDIV          = C.PPC_INS_FDIV
@@ -566,12 +873,25 @@ const (
 	PPC_INS_FSQRTS        = C.PPC_INS_FSQRTS
 	PPC_INS_FSUB          = C.PPC_INS_FSUB
 	PPC_INS_FSUBS         = C.PPC_INS_FSUBS
+	PPC_INS_FTDIV         = C.PPC_INS_FTDIV
+	PPC_INS_FTSQRT        = C.PPC_INS_FTSQRT
+	PPC_INS_HRFID         = C.PPC_INS_HRFID
 	PPC_INS_ICBI          = C.PPC_INS_ICBI
+	PPC_INS_ICBIEP        = C.PPC_INS_ICBIEP
+	PPC_INS_ICBLC         = C.PPC_INS_ICBLC
+	PPC_INS_ICBLQ         = C.PPC_INS_ICBLQ
 	PPC_INS_ICBT          = C.PPC_INS_ICBT
+	PPC_INS_ICBTLS        = C.PPC_INS_ICBTLS
 	PPC_INS_ICCCI         = C.PPC_INS_ICCCI
+	PPC_INS_ICI           = C.PPC_INS_ICI
+	PPC_INS_INSLWI        = C.PPC_INS_INSLWI
+	PPC_INS_INSRDI        = C.PPC_INS_INSRDI
+	PPC_INS_INSRWI        = C.PPC_INS_INSRWI
 	PPC_INS_ISEL          = C.PPC_INS_ISEL
 	PPC_INS_ISYNC         = C.PPC_INS_ISYNC
 	PPC_INS_LA            = C.PPC_INS_LA
+	PPC_INS_LBARX         = C.PPC_INS_LBARX
+	PPC_INS_LBEPX         = C.PPC_INS_LBEPX
 	PPC_INS_LBZ           = C.PPC_INS_LBZ
 	PPC_INS_LBZCIX        = C.PPC_INS_LBZCIX
 	PPC_INS_LBZU          = C.PPC_INS_LBZU
@@ -579,12 +899,15 @@ const (
 	PPC_INS_LBZX          = C.PPC_INS_LBZX
 	PPC_INS_LD            = C.PPC_INS_LD
 	PPC_INS_LDARX         = C.PPC_INS_LDARX
+	PPC_INS_LDAT          = C.PPC_INS_LDAT
 	PPC_INS_LDBRX         = C.PPC_INS_LDBRX
 	PPC_INS_LDCIX         = C.PPC_INS_LDCIX
+	PPC_INS_LDMX          = C.PPC_INS_LDMX
 	PPC_INS_LDU           = C.PPC_INS_LDU
 	PPC_INS_LDUX          = C.PPC_INS_LDUX
 	PPC_INS_LDX           = C.PPC_INS_LDX
 	PPC_INS_LFD           = C.PPC_INS_LFD
+	PPC_INS_LFDEPX        = C.PPC_INS_LFDEPX
 	PPC_INS_LFDU          = C.PPC_INS_LFDU
 	PPC_INS_LFDUX         = C.PPC_INS_LFDUX
 	PPC_INS_LFDX          = C.PPC_INS_LFDX
@@ -595,10 +918,12 @@ const (
 	PPC_INS_LFSUX         = C.PPC_INS_LFSUX
 	PPC_INS_LFSX          = C.PPC_INS_LFSX
 	PPC_INS_LHA           = C.PPC_INS_LHA
+	PPC_INS_LHARX         = C.PPC_INS_LHARX
 	PPC_INS_LHAU          = C.PPC_INS_LHAU
 	PPC_INS_LHAUX         = C.PPC_INS_LHAUX
 	PPC_INS_LHAX          = C.PPC_INS_LHAX
 	PPC_INS_LHBRX         = C.PPC_INS_LHBRX
+	PPC_INS_LHEPX         = C.PPC_INS_LHEPX
 	PPC_INS_LHZ           = C.PPC_INS_LHZ
 	PPC_INS_LHZCIX        = C.PPC_INS_LHZCIX
 	PPC_INS_LHZU          = C.PPC_INS_LHZU
@@ -607,6 +932,7 @@ const (
 	PPC_INS_LI            = C.PPC_INS_LI
 	PPC_INS_LIS           = C.PPC_INS_LIS
 	PPC_INS_LMW           = C.PPC_INS_LMW
+	PPC_INS_LNIA          = C.PPC_INS_LNIA
 	PPC_INS_LSWI          = C.PPC_INS_LSWI
 	PPC_INS_LVEBX         = C.PPC_INS_LVEBX
 	PPC_INS_LVEHX         = C.PPC_INS_LVEHX
@@ -617,49 +943,191 @@ const (
 	PPC_INS_LVXL          = C.PPC_INS_LVXL
 	PPC_INS_LWA           = C.PPC_INS_LWA
 	PPC_INS_LWARX         = C.PPC_INS_LWARX
+	PPC_INS_LWAT          = C.PPC_INS_LWAT
 	PPC_INS_LWAUX         = C.PPC_INS_LWAUX
 	PPC_INS_LWAX          = C.PPC_INS_LWAX
 	PPC_INS_LWBRX         = C.PPC_INS_LWBRX
+	PPC_INS_LWEPX         = C.PPC_INS_LWEPX
+	PPC_INS_LWSYNC        = C.PPC_INS_LWSYNC
 	PPC_INS_LWZ           = C.PPC_INS_LWZ
 	PPC_INS_LWZCIX        = C.PPC_INS_LWZCIX
 	PPC_INS_LWZU          = C.PPC_INS_LWZU
 	PPC_INS_LWZUX         = C.PPC_INS_LWZUX
 	PPC_INS_LWZX          = C.PPC_INS_LWZX
+	PPC_INS_LXSD          = C.PPC_INS_LXSD
 	PPC_INS_LXSDX         = C.PPC_INS_LXSDX
+	PPC_INS_LXSIBZX       = C.PPC_INS_LXSIBZX
+	PPC_INS_LXSIHZX       = C.PPC_INS_LXSIHZX
+	PPC_INS_LXSIWAX       = C.PPC_INS_LXSIWAX
+	PPC_INS_LXSIWZX       = C.PPC_INS_LXSIWZX
+	PPC_INS_LXSSP         = C.PPC_INS_LXSSP
+	PPC_INS_LXSSPX        = C.PPC_INS_LXSSPX
+	PPC_INS_LXV           = C.PPC_INS_LXV
+	PPC_INS_LXVB16X       = C.PPC_INS_LXVB16X
 	PPC_INS_LXVD2X        = C.PPC_INS_LXVD2X
 	PPC_INS_LXVDSX        = C.PPC_INS_LXVDSX
+	PPC_INS_LXVH8X        = C.PPC_INS_LXVH8X
+	PPC_INS_LXVL          = C.PPC_INS_LXVL
+	PPC_INS_LXVLL         = C.PPC_INS_LXVLL
 	PPC_INS_LXVW4X        = C.PPC_INS_LXVW4X
+	PPC_INS_LXVWSX        = C.PPC_INS_LXVWSX
+	PPC_INS_LXVX          = C.PPC_INS_LXVX
+	PPC_INS_MADDHD        = C.PPC_INS_MADDHD
+	PPC_INS_MADDHDU       = C.PPC_INS_MADDHDU
+	PPC_INS_MADDLD        = C.PPC_INS_MADDLD
 	PPC_INS_MBAR          = C.PPC_INS_MBAR
 	PPC_INS_MCRF          = C.PPC_INS_MCRF
 	PPC_INS_MCRFS         = C.PPC_INS_MCRFS
+	PPC_INS_MCRXRX        = C.PPC_INS_MCRXRX
+	PPC_INS_MFAMR         = C.PPC_INS_MFAMR
+	PPC_INS_MFASR         = C.PPC_INS_MFASR
+	PPC_INS_MFBHRBE       = C.PPC_INS_MFBHRBE
+	PPC_INS_MFBR0         = C.PPC_INS_MFBR0
+	PPC_INS_MFBR1         = C.PPC_INS_MFBR1
+	PPC_INS_MFBR2         = C.PPC_INS_MFBR2
+	PPC_INS_MFBR3         = C.PPC_INS_MFBR3
+	PPC_INS_MFBR4         = C.PPC_INS_MFBR4
+	PPC_INS_MFBR5         = C.PPC_INS_MFBR5
+	PPC_INS_MFBR6         = C.PPC_INS_MFBR6
+	PPC_INS_MFBR7         = C.PPC_INS_MFBR7
+	PPC_INS_MFCFAR        = C.PPC_INS_MFCFAR
 	PPC_INS_MFCR          = C.PPC_INS_MFCR
 	PPC_INS_MFCTR         = C.PPC_INS_MFCTR
+	PPC_INS_MFDAR         = C.PPC_INS_MFDAR
+	PPC_INS_MFDBATL       = C.PPC_INS_MFDBATL
+	PPC_INS_MFDBATU       = C.PPC_INS_MFDBATU
+	PPC_INS_MFDCCR        = C.PPC_INS_MFDCCR
 	PPC_INS_MFDCR         = C.PPC_INS_MFDCR
+	PPC_INS_MFDEAR        = C.PPC_INS_MFDEAR
+	PPC_INS_MFDEC         = C.PPC_INS_MFDEC
+	PPC_INS_MFDSCR        = C.PPC_INS_MFDSCR
+	PPC_INS_MFDSISR       = C.PPC_INS_MFDSISR
+	PPC_INS_MFESR         = C.PPC_INS_MFESR
+	PPC_INS_MFFPRD        = C.PPC_INS_MFFPRD
 	PPC_INS_MFFS          = C.PPC_INS_MFFS
+	PPC_INS_MFFSCDRN      = C.PPC_INS_MFFSCDRN
+	PPC_INS_MFFSCDRNI     = C.PPC_INS_MFFSCDRNI
+	PPC_INS_MFFSCE        = C.PPC_INS_MFFSCE
+	PPC_INS_MFFSCRN       = C.PPC_INS_MFFSCRN
+	PPC_INS_MFFSCRNI      = C.PPC_INS_MFFSCRNI
+	PPC_INS_MFFSL         = C.PPC_INS_MFFSL
+	PPC_INS_MFIBATL       = C.PPC_INS_MFIBATL
+	PPC_INS_MFIBATU       = C.PPC_INS_MFIBATU
+	PPC_INS_MFICCR        = C.PPC_INS_MFICCR
 	PPC_INS_MFLR          = C.PPC_INS_MFLR
 	PPC_INS_MFMSR         = C.PPC_INS_MFMSR
 	PPC_INS_MFOCRF        = C.PPC_INS_MFOCRF
+	PPC_INS_MFPID         = C.PPC_INS_MFPID
+	PPC_INS_MFPMR         = C.PPC_INS_MFPMR
+	PPC_INS_MFPVR         = C.PPC_INS_MFPVR
+	PPC_INS_MFRTCL        = C.PPC_INS_MFRTCL
+	PPC_INS_MFRTCU        = C.PPC_INS_MFRTCU
+	PPC_INS_MFSDR1        = C.PPC_INS_MFSDR1
+	PPC_INS_MFSPEFSCR     = C.PPC_INS_MFSPEFSCR
 	PPC_INS_MFSPR         = C.PPC_INS_MFSPR
+	PPC_INS_MFSPRG        = C.PPC_INS_MFSPRG
+	PPC_INS_MFSPRG0       = C.PPC_INS_MFSPRG0
+	PPC_INS_MFSPRG1       = C.PPC_INS_MFSPRG1
+	PPC_INS_MFSPRG2       = C.PPC_INS_MFSPRG2
+	PPC_INS_MFSPRG3       = C.PPC_INS_MFSPRG3
+	PPC_INS_MFSPRG4       = C.PPC_INS_MFSPRG4
+	PPC_INS_MFSPRG5       = C.PPC_INS_MFSPRG5
+	PPC_INS_MFSPRG6       = C.PPC_INS_MFSPRG6
+	PPC_INS_MFSPRG7       = C.PPC_INS_MFSPRG7
 	PPC_INS_MFSR          = C.PPC_INS_MFSR
 	PPC_INS_MFSRIN        = C.PPC_INS_MFSRIN
+	PPC_INS_MFSRR0        = C.PPC_INS_MFSRR0
+	PPC_INS_MFSRR1        = C.PPC_INS_MFSRR1
+	PPC_INS_MFSRR2        = C.PPC_INS_MFSRR2
+	PPC_INS_MFSRR3        = C.PPC_INS_MFSRR3
 	PPC_INS_MFTB          = C.PPC_INS_MFTB
+	PPC_INS_MFTBHI        = C.PPC_INS_MFTBHI
+	PPC_INS_MFTBL         = C.PPC_INS_MFTBL
+	PPC_INS_MFTBLO        = C.PPC_INS_MFTBLO
+	PPC_INS_MFTBU         = C.PPC_INS_MFTBU
+	PPC_INS_MFTCR         = C.PPC_INS_MFTCR
+	PPC_INS_MFVRD         = C.PPC_INS_MFVRD
+	PPC_INS_MFVRSAVE      = C.PPC_INS_MFVRSAVE
 	PPC_INS_MFVSCR        = C.PPC_INS_MFVSCR
+	PPC_INS_MFVSRD        = C.PPC_INS_MFVSRD
+	PPC_INS_MFVSRLD       = C.PPC_INS_MFVSRLD
+	PPC_INS_MFVSRWZ       = C.PPC_INS_MFVSRWZ
+	PPC_INS_MFXER         = C.PPC_INS_MFXER
+	PPC_INS_MODSD         = C.PPC_INS_MODSD
+	PPC_INS_MODSW         = C.PPC_INS_MODSW
+	PPC_INS_MODUD         = C.PPC_INS_MODUD
+	PPC_INS_MODUW         = C.PPC_INS_MODUW
+	PPC_INS_MR            = C.PPC_INS_MR
+	PPC_INS_MSGSYNC       = C.PPC_INS_MSGSYNC
 	PPC_INS_MSYNC         = C.PPC_INS_MSYNC
+	PPC_INS_MTAMR         = C.PPC_INS_MTAMR
+	PPC_INS_MTASR         = C.PPC_INS_MTASR
+	PPC_INS_MTBR0         = C.PPC_INS_MTBR0
+	PPC_INS_MTBR1         = C.PPC_INS_MTBR1
+	PPC_INS_MTBR2         = C.PPC_INS_MTBR2
+	PPC_INS_MTBR3         = C.PPC_INS_MTBR3
+	PPC_INS_MTBR4         = C.PPC_INS_MTBR4
+	PPC_INS_MTBR5         = C.PPC_INS_MTBR5
+	PPC_INS_MTBR6         = C.PPC_INS_MTBR6
+	PPC_INS_MTBR7         = C.PPC_INS_MTBR7
+	PPC_INS_MTCFAR        = C.PPC_INS_MTCFAR
+	PPC_INS_MTCR          = C.PPC_INS_MTCR
 	PPC_INS_MTCRF         = C.PPC_INS_MTCRF
 	PPC_INS_MTCTR         = C.PPC_INS_MTCTR
+	PPC_INS_MTDAR         = C.PPC_INS_MTDAR
+	PPC_INS_MTDBATL       = C.PPC_INS_MTDBATL
+	PPC_INS_MTDBATU       = C.PPC_INS_MTDBATU
+	PPC_INS_MTDCCR        = C.PPC_INS_MTDCCR
 	PPC_INS_MTDCR         = C.PPC_INS_MTDCR
+	PPC_INS_MTDEAR        = C.PPC_INS_MTDEAR
+	PPC_INS_MTDEC         = C.PPC_INS_MTDEC
+	PPC_INS_MTDSCR        = C.PPC_INS_MTDSCR
+	PPC_INS_MTDSISR       = C.PPC_INS_MTDSISR
+	PPC_INS_MTESR         = C.PPC_INS_MTESR
 	PPC_INS_MTFSB0        = C.PPC_INS_MTFSB0
 	PPC_INS_MTFSB1        = C.PPC_INS_MTFSB1
 	PPC_INS_MTFSF         = C.PPC_INS_MTFSF
 	PPC_INS_MTFSFI        = C.PPC_INS_MTFSFI
+	PPC_INS_MTIBATL       = C.PPC_INS_MTIBATL
+	PPC_INS_MTIBATU       = C.PPC_INS_MTIBATU
+	PPC_INS_MTICCR        = C.PPC_INS_MTICCR
 	PPC_INS_MTLR          = C.PPC_INS_MTLR
 	PPC_INS_MTMSR         = C.PPC_INS_MTMSR
 	PPC_INS_MTMSRD        = C.PPC_INS_MTMSRD
 	PPC_INS_MTOCRF        = C.PPC_INS_MTOCRF
+	PPC_INS_MTPID         = C.PPC_INS_MTPID
+	PPC_INS_MTPMR         = C.PPC_INS_MTPMR
+	PPC_INS_MTSDR1        = C.PPC_INS_MTSDR1
+	PPC_INS_MTSPEFSCR     = C.PPC_INS_MTSPEFSCR
 	PPC_INS_MTSPR         = C.PPC_INS_MTSPR
+	PPC_INS_MTSPRG        = C.PPC_INS_MTSPRG
+	PPC_INS_MTSPRG0       = C.PPC_INS_MTSPRG0
+	PPC_INS_MTSPRG1       = C.PPC_INS_MTSPRG1
+	PPC_INS_MTSPRG2       = C.PPC_INS_MTSPRG2
+	PPC_INS_MTSPRG3       = C.PPC_INS_MTSPRG3
+	PPC_INS_MTSPRG4       = C.PPC_INS_MTSPRG4
+	PPC_INS_MTSPRG5       = C.PPC_INS_MTSPRG5
+	PPC_INS_MTSPRG6       = C.PPC_INS_MTSPRG6
+	PPC_INS_MTSPRG7       = C.PPC_INS_MTSPRG7
 	PPC_INS_MTSR          = C.PPC_INS_MTSR
 	PPC_INS_MTSRIN        = C.PPC_INS_MTSRIN
+	PPC_INS_MTSRR0        = C.PPC_INS_MTSRR0
+	PPC_INS_MTSRR1        = C.PPC_INS_MTSRR1
+	PPC_INS_MTSRR2        = C.PPC_INS_MTSRR2
+	PPC_INS_MTSRR3        = C.PPC_INS_MTSRR3
+	PPC_INS_MTTBHI        = C.PPC_INS_MTTBHI
+	PPC_INS_MTTBL         = C.PPC_INS_MTTBL
+	PPC_INS_MTTBLO        = C.PPC_INS_MTTBLO
+	PPC_INS_MTTBU         = C.PPC_INS_MTTBU
+	PPC_INS_MTTCR         = C.PPC_INS_MTTCR
+	PPC_INS_MTVRSAVE      = C.PPC_INS_MTVRSAVE
 	PPC_INS_MTVSCR        = C.PPC_INS_MTVSCR
+	PPC_INS_MTVSRD        = C.PPC_INS_MTVSRD
+	PPC_INS_MTVSRDD       = C.PPC_INS_MTVSRDD
+	PPC_INS_MTVSRWA       = C.PPC_INS_MTVSRWA
+	PPC_INS_MTVSRWS       = C.PPC_INS_MTVSRWS
+	PPC_INS_MTVSRWZ       = C.PPC_INS_MTVSRWZ
+	PPC_INS_MTXER         = C.PPC_INS_MTXER
 	PPC_INS_MULHD         = C.PPC_INS_MULHD
 	PPC_INS_MULHDU        = C.PPC_INS_MULHDU
 	PPC_INS_MULHW         = C.PPC_INS_MULHW
@@ -668,28 +1136,38 @@ const (
 	PPC_INS_MULLI         = C.PPC_INS_MULLI
 	PPC_INS_MULLW         = C.PPC_INS_MULLW
 	PPC_INS_NAND          = C.PPC_INS_NAND
+	PPC_INS_NAP           = C.PPC_INS_NAP
 	PPC_INS_NEG           = C.PPC_INS_NEG
 	PPC_INS_NOP           = C.PPC_INS_NOP
-	PPC_INS_ORI           = C.PPC_INS_ORI
 	PPC_INS_NOR           = C.PPC_INS_NOR
+	PPC_INS_NOT           = C.PPC_INS_NOT
 	PPC_INS_OR            = C.PPC_INS_OR
 	PPC_INS_ORC           = C.PPC_INS_ORC
+	PPC_INS_ORI           = C.PPC_INS_ORI
 	PPC_INS_ORIS          = C.PPC_INS_ORIS
+	PPC_INS_PASTE         = C.PPC_INS_PASTE
+	PPC_INS_PASTE_LAST    = C.PPC_INS_PASTE_LAST
+	PPC_INS_POPCNTB       = C.PPC_INS_POPCNTB
 	PPC_INS_POPCNTD       = C.PPC_INS_POPCNTD
 	PPC_INS_POPCNTW       = C.PPC_INS_POPCNTW
+	PPC_INS_PTESYNC       = C.PPC_INS_PTESYNC
 	PPC_INS_QVALIGNI      = C.PPC_INS_QVALIGNI
 	PPC_INS_QVESPLATI     = C.PPC_INS_QVESPLATI
 	PPC_INS_QVFABS        = C.PPC_INS_QVFABS
 	PPC_INS_QVFADD        = C.PPC_INS_QVFADD
 	PPC_INS_QVFADDS       = C.PPC_INS_QVFADDS
+	PPC_INS_QVFAND        = C.PPC_INS_QVFAND
+	PPC_INS_QVFANDC       = C.PPC_INS_QVFANDC
 	PPC_INS_QVFCFID       = C.PPC_INS_QVFCFID
 	PPC_INS_QVFCFIDS      = C.PPC_INS_QVFCFIDS
 	PPC_INS_QVFCFIDU      = C.PPC_INS_QVFCFIDU
 	PPC_INS_QVFCFIDUS     = C.PPC_INS_QVFCFIDUS
+	PPC_INS_QVFCLR        = C.PPC_INS_QVFCLR
 	PPC_INS_QVFCMPEQ      = C.PPC_INS_QVFCMPEQ
 	PPC_INS_QVFCMPGT      = C.PPC_INS_QVFCMPGT
 	PPC_INS_QVFCMPLT      = C.PPC_INS_QVFCMPLT
 	PPC_INS_QVFCPSGN      = C.PPC_INS_QVFCPSGN
+	PPC_INS_QVFCTFB       = C.PPC_INS_QVFCTFB
 	PPC_INS_QVFCTID       = C.PPC_INS_QVFCTID
 	PPC_INS_QVFCTIDU      = C.PPC_INS_QVFCTIDU
 	PPC_INS_QVFCTIDUZ     = C.PPC_INS_QVFCTIDUZ
@@ -698,6 +1176,7 @@ const (
 	PPC_INS_QVFCTIWU      = C.PPC_INS_QVFCTIWU
 	PPC_INS_QVFCTIWUZ     = C.PPC_INS_QVFCTIWUZ
 	PPC_INS_QVFCTIWZ      = C.PPC_INS_QVFCTIWZ
+	PPC_INS_QVFEQU        = C.PPC_INS_QVFEQU
 	PPC_INS_QVFLOGICAL    = C.PPC_INS_QVFLOGICAL
 	PPC_INS_QVFMADD       = C.PPC_INS_QVFMADD
 	PPC_INS_QVFMADDS      = C.PPC_INS_QVFMADDS
@@ -707,11 +1186,16 @@ const (
 	PPC_INS_QVFMUL        = C.PPC_INS_QVFMUL
 	PPC_INS_QVFMULS       = C.PPC_INS_QVFMULS
 	PPC_INS_QVFNABS       = C.PPC_INS_QVFNABS
+	PPC_INS_QVFNAND       = C.PPC_INS_QVFNAND
 	PPC_INS_QVFNEG        = C.PPC_INS_QVFNEG
 	PPC_INS_QVFNMADD      = C.PPC_INS_QVFNMADD
 	PPC_INS_QVFNMADDS     = C.PPC_INS_QVFNMADDS
 	PPC_INS_QVFNMSUB      = C.PPC_INS_QVFNMSUB
 	PPC_INS_QVFNMSUBS     = C.PPC_INS_QVFNMSUBS
+	PPC_INS_QVFNOR        = C.PPC_INS_QVFNOR
+	PPC_INS_QVFNOT        = C.PPC_INS_QVFNOT
+	PPC_INS_QVFOR         = C.PPC_INS_QVFOR
+	PPC_INS_QVFORC        = C.PPC_INS_QVFORC
 	PPC_INS_QVFPERM       = C.PPC_INS_QVFPERM
 	PPC_INS_QVFRE         = C.PPC_INS_QVFRE
 	PPC_INS_QVFRES        = C.PPC_INS_QVFRES
@@ -723,6 +1207,7 @@ const (
 	PPC_INS_QVFRSQRTE     = C.PPC_INS_QVFRSQRTE
 	PPC_INS_QVFRSQRTES    = C.PPC_INS_QVFRSQRTES
 	PPC_INS_QVFSEL        = C.PPC_INS_QVFSEL
+	PPC_INS_QVFSET        = C.PPC_INS_QVFSET
 	PPC_INS_QVFSUB        = C.PPC_INS_QVFSUB
 	PPC_INS_QVFSUBS       = C.PPC_INS_QVFSUBS
 	PPC_INS_QVFTSTNAN     = C.PPC_INS_QVFTSTNAN
@@ -730,6 +1215,7 @@ const (
 	PPC_INS_QVFXMADDS     = C.PPC_INS_QVFXMADDS
 	PPC_INS_QVFXMUL       = C.PPC_INS_QVFXMUL
 	PPC_INS_QVFXMULS      = C.PPC_INS_QVFXMULS
+	PPC_INS_QVFXOR        = C.PPC_INS_QVFXOR
 	PPC_INS_QVFXXCPNMADD  = C.PPC_INS_QVFXXCPNMADD
 	PPC_INS_QVFXXCPNMADDS = C.PPC_INS_QVFXXCPNMADDS
 	PPC_INS_QVFXXMADD     = C.PPC_INS_QVFXXMADD
@@ -797,6 +1283,7 @@ const (
 	PPC_INS_QVSTFSXIA     = C.PPC_INS_QVSTFSXIA
 	PPC_INS_RFCI          = C.PPC_INS_RFCI
 	PPC_INS_RFDI          = C.PPC_INS_RFDI
+	PPC_INS_RFEBB         = C.PPC_INS_RFEBB
 	PPC_INS_RFI           = C.PPC_INS_RFI
 	PPC_INS_RFID          = C.PPC_INS_RFID
 	PPC_INS_RFMCI         = C.PPC_INS_RFMCI
@@ -809,25 +1296,42 @@ const (
 	PPC_INS_RLWIMI        = C.PPC_INS_RLWIMI
 	PPC_INS_RLWINM        = C.PPC_INS_RLWINM
 	PPC_INS_RLWNM         = C.PPC_INS_RLWNM
+	PPC_INS_ROTLD         = C.PPC_INS_ROTLD
+	PPC_INS_ROTLDI        = C.PPC_INS_ROTLDI
+	PPC_INS_ROTLW         = C.PPC_INS_ROTLW
+	PPC_INS_ROTLWI        = C.PPC_INS_ROTLWI
+	PPC_INS_ROTRDI        = C.PPC_INS_ROTRDI
+	PPC_INS_ROTRWI        = C.PPC_INS_ROTRWI
 	PPC_INS_SC            = C.PPC_INS_SC
+	PPC_INS_SETB          = C.PPC_INS_SETB
 	PPC_INS_SLBIA         = C.PPC_INS_SLBIA
 	PPC_INS_SLBIE         = C.PPC_INS_SLBIE
+	PPC_INS_SLBIEG        = C.PPC_INS_SLBIEG
 	PPC_INS_SLBMFEE       = C.PPC_INS_SLBMFEE
+	PPC_INS_SLBMFEV       = C.PPC_INS_SLBMFEV
 	PPC_INS_SLBMTE        = C.PPC_INS_SLBMTE
+	PPC_INS_SLBSYNC       = C.PPC_INS_SLBSYNC
 	PPC_INS_SLD           = C.PPC_INS_SLD
+	PPC_INS_SLDI          = C.PPC_INS_SLDI
 	PPC_INS_SLW           = C.PPC_INS_SLW
+	PPC_INS_SLWI          = C.PPC_INS_SLWI
 	PPC_INS_SRAD          = C.PPC_INS_SRAD
 	PPC_INS_SRADI         = C.PPC_INS_SRADI
 	PPC_INS_SRAW          = C.PPC_INS_SRAW
 	PPC_INS_SRAWI         = C.PPC_INS_SRAWI
 	PPC_INS_SRD           = C.PPC_INS_SRD
+	PPC_INS_SRDI          = C.PPC_INS_SRDI
 	PPC_INS_SRW           = C.PPC_INS_SRW
+	PPC_INS_SRWI          = C.PPC_INS_SRWI
 	PPC_INS_STB           = C.PPC_INS_STB
 	PPC_INS_STBCIX        = C.PPC_INS_STBCIX
+	PPC_INS_STBCX         = C.PPC_INS_STBCX
+	PPC_INS_STBEPX        = C.PPC_INS_STBEPX
 	PPC_INS_STBU          = C.PPC_INS_STBU
 	PPC_INS_STBUX         = C.PPC_INS_STBUX
 	PPC_INS_STBX          = C.PPC_INS_STBX
 	PPC_INS_STD           = C.PPC_INS_STD
+	PPC_INS_STDAT         = C.PPC_INS_STDAT
 	PPC_INS_STDBRX        = C.PPC_INS_STDBRX
 	PPC_INS_STDCIX        = C.PPC_INS_STDCIX
 	PPC_INS_STDCX         = C.PPC_INS_STDCX
@@ -835,6 +1339,7 @@ const (
 	PPC_INS_STDUX         = C.PPC_INS_STDUX
 	PPC_INS_STDX          = C.PPC_INS_STDX
 	PPC_INS_STFD          = C.PPC_INS_STFD
+	PPC_INS_STFDEPX       = C.PPC_INS_STFDEPX
 	PPC_INS_STFDU         = C.PPC_INS_STFDU
 	PPC_INS_STFDUX        = C.PPC_INS_STFDUX
 	PPC_INS_STFDX         = C.PPC_INS_STFDX
@@ -846,10 +1351,13 @@ const (
 	PPC_INS_STH           = C.PPC_INS_STH
 	PPC_INS_STHBRX        = C.PPC_INS_STHBRX
 	PPC_INS_STHCIX        = C.PPC_INS_STHCIX
+	PPC_INS_STHCX         = C.PPC_INS_STHCX
+	PPC_INS_STHEPX        = C.PPC_INS_STHEPX
 	PPC_INS_STHU          = C.PPC_INS_STHU
 	PPC_INS_STHUX         = C.PPC_INS_STHUX
 	PPC_INS_STHX          = C.PPC_INS_STHX
 	PPC_INS_STMW          = C.PPC_INS_STMW
+	PPC_INS_STOP          = C.PPC_INS_STOP
 	PPC_INS_STSWI         = C.PPC_INS_STSWI
 	PPC_INS_STVEBX        = C.PPC_INS_STVEBX
 	PPC_INS_STVEHX        = C.PPC_INS_STVEHX
@@ -857,24 +1365,82 @@ const (
 	PPC_INS_STVX          = C.PPC_INS_STVX
 	PPC_INS_STVXL         = C.PPC_INS_STVXL
 	PPC_INS_STW           = C.PPC_INS_STW
+	PPC_INS_STWAT         = C.PPC_INS_STWAT
 	PPC_INS_STWBRX        = C.PPC_INS_STWBRX
 	PPC_INS_STWCIX        = C.PPC_INS_STWCIX
 	PPC_INS_STWCX         = C.PPC_INS_STWCX
+	PPC_INS_STWEPX        = C.PPC_INS_STWEPX
 	PPC_INS_STWU          = C.PPC_INS_STWU
 	PPC_INS_STWUX         = C.PPC_INS_STWUX
 	PPC_INS_STWX          = C.PPC_INS_STWX
+	PPC_INS_STXSD         = C.PPC_INS_STXSD
 	PPC_INS_STXSDX        = C.PPC_INS_STXSDX
+	PPC_INS_STXSIBX       = C.PPC_INS_STXSIBX
+	PPC_INS_STXSIHX       = C.PPC_INS_STXSIHX
+	PPC_INS_STXSIWX       = C.PPC_INS_STXSIWX
+	PPC_INS_STXSSP        = C.PPC_INS_STXSSP
+	PPC_INS_STXSSPX       = C.PPC_INS_STXSSPX
+	PPC_INS_STXV          = C.PPC_INS_STXV
+	PPC_INS_STXVB16X      = C.PPC_INS_STXVB16X
 	PPC_INS_STXVD2X       = C.PPC_INS_STXVD2X
+	PPC_INS_STXVH8X       = C.PPC_INS_STXVH8X
+	PPC_INS_STXVL         = C.PPC_INS_STXVL
+	PPC_INS_STXVLL        = C.PPC_INS_STXVLL
 	PPC_INS_STXVW4X       = C.PPC_INS_STXVW4X
+	PPC_INS_STXVX         = C.PPC_INS_STXVX
+	PPC_INS_SUB           = C.PPC_INS_SUB
+	PPC_INS_SUBC          = C.PPC_INS_SUBC
 	PPC_INS_SUBF          = C.PPC_INS_SUBF
 	PPC_INS_SUBFC         = C.PPC_INS_SUBFC
 	PPC_INS_SUBFE         = C.PPC_INS_SUBFE
 	PPC_INS_SUBFIC        = C.PPC_INS_SUBFIC
 	PPC_INS_SUBFME        = C.PPC_INS_SUBFME
 	PPC_INS_SUBFZE        = C.PPC_INS_SUBFZE
+	PPC_INS_SUBI          = C.PPC_INS_SUBI
+	PPC_INS_SUBIC         = C.PPC_INS_SUBIC
+	PPC_INS_SUBIS         = C.PPC_INS_SUBIS
+	PPC_INS_SUBPCIS       = C.PPC_INS_SUBPCIS
 	PPC_INS_SYNC          = C.PPC_INS_SYNC
+	PPC_INS_TABORT        = C.PPC_INS_TABORT
+	PPC_INS_TABORTDC      = C.PPC_INS_TABORTDC
+	PPC_INS_TABORTDCI     = C.PPC_INS_TABORTDCI
+	PPC_INS_TABORTWC      = C.PPC_INS_TABORTWC
+	PPC_INS_TABORTWCI     = C.PPC_INS_TABORTWCI
+	PPC_INS_TBEGIN        = C.PPC_INS_TBEGIN
+	PPC_INS_TCHECK        = C.PPC_INS_TCHECK
 	PPC_INS_TD            = C.PPC_INS_TD
+	PPC_INS_TDEQ          = C.PPC_INS_TDEQ
+	PPC_INS_TDEQI         = C.PPC_INS_TDEQI
+	PPC_INS_TDGE          = C.PPC_INS_TDGE
+	PPC_INS_TDGEI         = C.PPC_INS_TDGEI
+	PPC_INS_TDGT          = C.PPC_INS_TDGT
+	PPC_INS_TDGTI         = C.PPC_INS_TDGTI
 	PPC_INS_TDI           = C.PPC_INS_TDI
+	PPC_INS_TDLE          = C.PPC_INS_TDLE
+	PPC_INS_TDLEI         = C.PPC_INS_TDLEI
+	PPC_INS_TDLGE         = C.PPC_INS_TDLGE
+	PPC_INS_TDLGEI        = C.PPC_INS_TDLGEI
+	PPC_INS_TDLGT         = C.PPC_INS_TDLGT
+	PPC_INS_TDLGTI        = C.PPC_INS_TDLGTI
+	PPC_INS_TDLLE         = C.PPC_INS_TDLLE
+	PPC_INS_TDLLEI        = C.PPC_INS_TDLLEI
+	PPC_INS_TDLLT         = C.PPC_INS_TDLLT
+	PPC_INS_TDLLTI        = C.PPC_INS_TDLLTI
+	PPC_INS_TDLNG         = C.PPC_INS_TDLNG
+	PPC_INS_TDLNGI        = C.PPC_INS_TDLNGI
+	PPC_INS_TDLNL         = C.PPC_INS_TDLNL
+	PPC_INS_TDLNLI        = C.PPC_INS_TDLNLI
+	PPC_INS_TDLT          = C.PPC_INS_TDLT
+	PPC_INS_TDLTI         = C.PPC_INS_TDLTI
+	PPC_INS_TDNE          = C.PPC_INS_TDNE
+	PPC_INS_TDNEI         = C.PPC_INS_TDNEI
+	PPC_INS_TDNG          = C.PPC_INS_TDNG
+	PPC_INS_TDNGI         = C.PPC_INS_TDNGI
+	PPC_INS_TDNL          = C.PPC_INS_TDNL
+	PPC_INS_TDNLI         = C.PPC_INS_TDNLI
+	PPC_INS_TDU           = C.PPC_INS_TDU
+	PPC_INS_TDUI          = C.PPC_INS_TDUI
+	PPC_INS_TEND          = C.PPC_INS_TEND
 	PPC_INS_TLBIA         = C.PPC_INS_TLBIA
 	PPC_INS_TLBIE         = C.PPC_INS_TLBIE
 	PPC_INS_TLBIEL        = C.PPC_INS_TLBIEL
@@ -882,13 +1448,56 @@ const (
 	PPC_INS_TLBLD         = C.PPC_INS_TLBLD
 	PPC_INS_TLBLI         = C.PPC_INS_TLBLI
 	PPC_INS_TLBRE         = C.PPC_INS_TLBRE
+	PPC_INS_TLBREHI       = C.PPC_INS_TLBREHI
+	PPC_INS_TLBRELO       = C.PPC_INS_TLBRELO
 	PPC_INS_TLBSX         = C.PPC_INS_TLBSX
 	PPC_INS_TLBSYNC       = C.PPC_INS_TLBSYNC
 	PPC_INS_TLBWE         = C.PPC_INS_TLBWE
+	PPC_INS_TLBWEHI       = C.PPC_INS_TLBWEHI
+	PPC_INS_TLBWELO       = C.PPC_INS_TLBWELO
 	PPC_INS_TRAP          = C.PPC_INS_TRAP
+	PPC_INS_TRECHKPT      = C.PPC_INS_TRECHKPT
+	PPC_INS_TRECLAIM      = C.PPC_INS_TRECLAIM
+	PPC_INS_TSR           = C.PPC_INS_TSR
 	PPC_INS_TW            = C.PPC_INS_TW
+	PPC_INS_TWEQ          = C.PPC_INS_TWEQ
+	PPC_INS_TWEQI         = C.PPC_INS_TWEQI
+	PPC_INS_TWGE          = C.PPC_INS_TWGE
+	PPC_INS_TWGEI         = C.PPC_INS_TWGEI
+	PPC_INS_TWGT          = C.PPC_INS_TWGT
+	PPC_INS_TWGTI         = C.PPC_INS_TWGTI
 	PPC_INS_TWI           = C.PPC_INS_TWI
+	PPC_INS_TWLE          = C.PPC_INS_TWLE
+	PPC_INS_TWLEI         = C.PPC_INS_TWLEI
+	PPC_INS_TWLGE         = C.PPC_INS_TWLGE
+	PPC_INS_TWLGEI        = C.PPC_INS_TWLGEI
+	PPC_INS_TWLGT         = C.PPC_INS_TWLGT
+	PPC_INS_TWLGTI        = C.PPC_INS_TWLGTI
+	PPC_INS_TWLLE         = C.PPC_INS_TWLLE
+	PPC_INS_TWLLEI        = C.PPC_INS_TWLLEI
+	PPC_INS_TWLLT         = C.PPC_INS_TWLLT
+	PPC_INS_TWLLTI        = C.PPC_INS_TWLLTI
+	PPC_INS_TWLNG         = C.PPC_INS_TWLNG
+	PPC_INS_TWLNGI        = C.PPC_INS_TWLNGI
+	PPC_INS_TWLNL         = C.PPC_INS_TWLNL
+	PPC_INS_TWLNLI        = C.PPC_INS_TWLNLI
+	PPC_INS_TWLT          = C.PPC_INS_TWLT
+	PPC_INS_TWLTI         = C.PPC_INS_TWLTI
+	PPC_INS_TWNE          = C.PPC_INS_TWNE
+	PPC_INS_TWNEI         = C.PPC_INS_TWNEI
+	PPC_INS_TWNG          = C.PPC_INS_TWNG
+	PPC_INS_TWNGI         = C.PPC_INS_TWNGI
+	PPC_INS_TWNL          = C.PPC_INS_TWNL
+	PPC_INS_TWNLI         = C.PPC_INS_TWNLI
+	PPC_INS_TWU           = C.PPC_INS_TWU
+	PPC_INS_TWUI          = C.PPC_INS_TWUI
+	PPC_INS_VABSDUB       = C.PPC_INS_VABSDUB
+	PPC_INS_VABSDUH       = C.PPC_INS_VABSDUH
+	PPC_INS_VABSDUW       = C.PPC_INS_VABSDUW
+	PPC_INS_VADDCUQ       = C.PPC_INS_VADDCUQ
 	PPC_INS_VADDCUW       = C.PPC_INS_VADDCUW
+	PPC_INS_VADDECUQ      = C.PPC_INS_VADDECUQ
+	PPC_INS_VADDEUQM      = C.PPC_INS_VADDEUQM
 	PPC_INS_VADDFP        = C.PPC_INS_VADDFP
 	PPC_INS_VADDSBS       = C.PPC_INS_VADDSBS
 	PPC_INS_VADDSHS       = C.PPC_INS_VADDSHS
@@ -898,6 +1507,7 @@ const (
 	PPC_INS_VADDUDM       = C.PPC_INS_VADDUDM
 	PPC_INS_VADDUHM       = C.PPC_INS_VADDUHM
 	PPC_INS_VADDUHS       = C.PPC_INS_VADDUHS
+	PPC_INS_VADDUQM       = C.PPC_INS_VADDUQM
 	PPC_INS_VADDUWM       = C.PPC_INS_VADDUWM
 	PPC_INS_VADDUWS       = C.PPC_INS_VADDUWS
 	PPC_INS_VAND          = C.PPC_INS_VAND
@@ -908,11 +1518,16 @@ const (
 	PPC_INS_VAVGUB        = C.PPC_INS_VAVGUB
 	PPC_INS_VAVGUH        = C.PPC_INS_VAVGUH
 	PPC_INS_VAVGUW        = C.PPC_INS_VAVGUW
+	PPC_INS_VBPERMD       = C.PPC_INS_VBPERMD
+	PPC_INS_VBPERMQ       = C.PPC_INS_VBPERMQ
 	PPC_INS_VCFSX         = C.PPC_INS_VCFSX
 	PPC_INS_VCFUX         = C.PPC_INS_VCFUX
+	PPC_INS_VCIPHER       = C.PPC_INS_VCIPHER
+	PPC_INS_VCIPHERLAST   = C.PPC_INS_VCIPHERLAST
 	PPC_INS_VCLZB         = C.PPC_INS_VCLZB
 	PPC_INS_VCLZD         = C.PPC_INS_VCLZD
 	PPC_INS_VCLZH         = C.PPC_INS_VCLZH
+	PPC_INS_VCLZLSBB      = C.PPC_INS_VCLZLSBB
 	PPC_INS_VCLZW         = C.PPC_INS_VCLZW
 	PPC_INS_VCMPBFP       = C.PPC_INS_VCMPBFP
 	PPC_INS_VCMPEQFP      = C.PPC_INS_VCMPEQFP
@@ -930,10 +1545,41 @@ const (
 	PPC_INS_VCMPGTUD      = C.PPC_INS_VCMPGTUD
 	PPC_INS_VCMPGTUH      = C.PPC_INS_VCMPGTUH
 	PPC_INS_VCMPGTUW      = C.PPC_INS_VCMPGTUW
+	PPC_INS_VCMPNEB       = C.PPC_INS_VCMPNEB
+	PPC_INS_VCMPNEH       = C.PPC_INS_VCMPNEH
+	PPC_INS_VCMPNEW       = C.PPC_INS_VCMPNEW
+	PPC_INS_VCMPNEZB      = C.PPC_INS_VCMPNEZB
+	PPC_INS_VCMPNEZH      = C.PPC_INS_VCMPNEZH
+	PPC_INS_VCMPNEZW      = C.PPC_INS_VCMPNEZW
 	PPC_INS_VCTSXS        = C.PPC_INS_VCTSXS
 	PPC_INS_VCTUXS        = C.PPC_INS_VCTUXS
+	PPC_INS_VCTZB         = C.PPC_INS_VCTZB
+	PPC_INS_VCTZD         = C.PPC_INS_VCTZD
+	PPC_INS_VCTZH         = C.PPC_INS_VCTZH
+	PPC_INS_VCTZLSBB      = C.PPC_INS_VCTZLSBB
+	PPC_INS_VCTZW         = C.PPC_INS_VCTZW
 	PPC_INS_VEQV          = C.PPC_INS_VEQV
 	PPC_INS_VEXPTEFP      = C.PPC_INS_VEXPTEFP
+	PPC_INS_VEXTRACTD     = C.PPC_INS_VEXTRACTD
+	PPC_INS_VEXTRACTUB    = C.PPC_INS_VEXTRACTUB
+	PPC_INS_VEXTRACTUH    = C.PPC_INS_VEXTRACTUH
+	PPC_INS_VEXTRACTUW    = C.PPC_INS_VEXTRACTUW
+	PPC_INS_VEXTSB2D      = C.PPC_INS_VEXTSB2D
+	PPC_INS_VEXTSB2W      = C.PPC_INS_VEXTSB2W
+	PPC_INS_VEXTSH2D      = C.PPC_INS_VEXTSH2D
+	PPC_INS_VEXTSH2W      = C.PPC_INS_VEXTSH2W
+	PPC_INS_VEXTSW2D      = C.PPC_INS_VEXTSW2D
+	PPC_INS_VEXTUBLX      = C.PPC_INS_VEXTUBLX
+	PPC_INS_VEXTUBRX      = C.PPC_INS_VEXTUBRX
+	PPC_INS_VEXTUHLX      = C.PPC_INS_VEXTUHLX
+	PPC_INS_VEXTUHRX      = C.PPC_INS_VEXTUHRX
+	PPC_INS_VEXTUWLX      = C.PPC_INS_VEXTUWLX
+	PPC_INS_VEXTUWRX      = C.PPC_INS_VEXTUWRX
+	PPC_INS_VGBBD         = C.PPC_INS_VGBBD
+	PPC_INS_VINSERTB      = C.PPC_INS_VINSERTB
+	PPC_INS_VINSERTD      = C.PPC_INS_VINSERTD
+	PPC_INS_VINSERTH      = C.PPC_INS_VINSERTH
+	PPC_INS_VINSERTW      = C.PPC_INS_VINSERTW
 	PPC_INS_VLOGEFP       = C.PPC_INS_VLOGEFP
 	PPC_INS_VMADDFP       = C.PPC_INS_VMADDFP
 	PPC_INS_VMAXFP        = C.PPC_INS_VMAXFP
@@ -947,28 +1593,35 @@ const (
 	PPC_INS_VMAXUW        = C.PPC_INS_VMAXUW
 	PPC_INS_VMHADDSHS     = C.PPC_INS_VMHADDSHS
 	PPC_INS_VMHRADDSHS    = C.PPC_INS_VMHRADDSHS
-	PPC_INS_VMINUD        = C.PPC_INS_VMINUD
 	PPC_INS_VMINFP        = C.PPC_INS_VMINFP
 	PPC_INS_VMINSB        = C.PPC_INS_VMINSB
 	PPC_INS_VMINSD        = C.PPC_INS_VMINSD
 	PPC_INS_VMINSH        = C.PPC_INS_VMINSH
 	PPC_INS_VMINSW        = C.PPC_INS_VMINSW
 	PPC_INS_VMINUB        = C.PPC_INS_VMINUB
+	PPC_INS_VMINUD        = C.PPC_INS_VMINUD
 	PPC_INS_VMINUH        = C.PPC_INS_VMINUH
 	PPC_INS_VMINUW        = C.PPC_INS_VMINUW
 	PPC_INS_VMLADDUHM     = C.PPC_INS_VMLADDUHM
+	PPC_INS_VMR           = C.PPC_INS_VMR
+	PPC_INS_VMRGEW        = C.PPC_INS_VMRGEW
 	PPC_INS_VMRGHB        = C.PPC_INS_VMRGHB
 	PPC_INS_VMRGHH        = C.PPC_INS_VMRGHH
 	PPC_INS_VMRGHW        = C.PPC_INS_VMRGHW
 	PPC_INS_VMRGLB        = C.PPC_INS_VMRGLB
 	PPC_INS_VMRGLH        = C.PPC_INS_VMRGLH
 	PPC_INS_VMRGLW        = C.PPC_INS_VMRGLW
+	PPC_INS_VMRGOW        = C.PPC_INS_VMRGOW
 	PPC_INS_VMSUMMBM      = C.PPC_INS_VMSUMMBM
 	PPC_INS_VMSUMSHM      = C.PPC_INS_VMSUMSHM
 	PPC_INS_VMSUMSHS      = C.PPC_INS_VMSUMSHS
 	PPC_INS_VMSUMUBM      = C.PPC_INS_VMSUMUBM
 	PPC_INS_VMSUMUHM      = C.PPC_INS_VMSUMUHM
 	PPC_INS_VMSUMUHS      = C.PPC_INS_VMSUMUHS
+	PPC_INS_VMUL10CUQ     = C.PPC_INS_VMUL10CUQ
+	PPC_INS_VMUL10ECUQ    = C.PPC_INS_VMUL10ECUQ
+	PPC_INS_VMUL10EUQ     = C.PPC_INS_VMUL10EUQ
+	PPC_INS_VMUL10UQ      = C.PPC_INS_VMUL10UQ
 	PPC_INS_VMULESB       = C.PPC_INS_VMULESB
 	PPC_INS_VMULESH       = C.PPC_INS_VMULESH
 	PPC_INS_VMULESW       = C.PPC_INS_VMULESW
@@ -983,24 +1636,42 @@ const (
 	PPC_INS_VMULOUW       = C.PPC_INS_VMULOUW
 	PPC_INS_VMULUWM       = C.PPC_INS_VMULUWM
 	PPC_INS_VNAND         = C.PPC_INS_VNAND
+	PPC_INS_VNCIPHER      = C.PPC_INS_VNCIPHER
+	PPC_INS_VNCIPHERLAST  = C.PPC_INS_VNCIPHERLAST
+	PPC_INS_VNEGD         = C.PPC_INS_VNEGD
+	PPC_INS_VNEGW         = C.PPC_INS_VNEGW
 	PPC_INS_VNMSUBFP      = C.PPC_INS_VNMSUBFP
 	PPC_INS_VNOR          = C.PPC_INS_VNOR
+	PPC_INS_VNOT          = C.PPC_INS_VNOT
 	PPC_INS_VOR           = C.PPC_INS_VOR
 	PPC_INS_VORC          = C.PPC_INS_VORC
 	PPC_INS_VPERM         = C.PPC_INS_VPERM
+	PPC_INS_VPERMR        = C.PPC_INS_VPERMR
+	PPC_INS_VPERMXOR      = C.PPC_INS_VPERMXOR
 	PPC_INS_VPKPX         = C.PPC_INS_VPKPX
+	PPC_INS_VPKSDSS       = C.PPC_INS_VPKSDSS
+	PPC_INS_VPKSDUS       = C.PPC_INS_VPKSDUS
 	PPC_INS_VPKSHSS       = C.PPC_INS_VPKSHSS
 	PPC_INS_VPKSHUS       = C.PPC_INS_VPKSHUS
 	PPC_INS_VPKSWSS       = C.PPC_INS_VPKSWSS
 	PPC_INS_VPKSWUS       = C.PPC_INS_VPKSWUS
+	PPC_INS_VPKUDUM       = C.PPC_INS_VPKUDUM
+	PPC_INS_VPKUDUS       = C.PPC_INS_VPKUDUS
 	PPC_INS_VPKUHUM       = C.PPC_INS_VPKUHUM
 	PPC_INS_VPKUHUS       = C.PPC_INS_VPKUHUS
 	PPC_INS_VPKUWUM       = C.PPC_INS_VPKUWUM
 	PPC_INS_VPKUWUS       = C.PPC_INS_VPKUWUS
+	PPC_INS_VPMSUMB       = C.PPC_INS_VPMSUMB
+	PPC_INS_VPMSUMD       = C.PPC_INS_VPMSUMD
+	PPC_INS_VPMSUMH       = C.PPC_INS_VPMSUMH
+	PPC_INS_VPMSUMW       = C.PPC_INS_VPMSUMW
 	PPC_INS_VPOPCNTB      = C.PPC_INS_VPOPCNTB
 	PPC_INS_VPOPCNTD      = C.PPC_INS_VPOPCNTD
 	PPC_INS_VPOPCNTH      = C.PPC_INS_VPOPCNTH
 	PPC_INS_VPOPCNTW      = C.PPC_INS_VPOPCNTW
+	PPC_INS_VPRTYBD       = C.PPC_INS_VPRTYBD
+	PPC_INS_VPRTYBQ       = C.PPC_INS_VPRTYBQ
+	PPC_INS_VPRTYBW       = C.PPC_INS_VPRTYBW
 	PPC_INS_VREFP         = C.PPC_INS_VREFP
 	PPC_INS_VRFIM         = C.PPC_INS_VRFIM
 	PPC_INS_VRFIN         = C.PPC_INS_VRFIN
@@ -1008,16 +1679,24 @@ const (
 	PPC_INS_VRFIZ         = C.PPC_INS_VRFIZ
 	PPC_INS_VRLB          = C.PPC_INS_VRLB
 	PPC_INS_VRLD          = C.PPC_INS_VRLD
+	PPC_INS_VRLDMI        = C.PPC_INS_VRLDMI
+	PPC_INS_VRLDNM        = C.PPC_INS_VRLDNM
 	PPC_INS_VRLH          = C.PPC_INS_VRLH
 	PPC_INS_VRLW          = C.PPC_INS_VRLW
+	PPC_INS_VRLWMI        = C.PPC_INS_VRLWMI
+	PPC_INS_VRLWNM        = C.PPC_INS_VRLWNM
 	PPC_INS_VRSQRTEFP     = C.PPC_INS_VRSQRTEFP
+	PPC_INS_VSBOX         = C.PPC_INS_VSBOX
 	PPC_INS_VSEL          = C.PPC_INS_VSEL
+	PPC_INS_VSHASIGMAD    = C.PPC_INS_VSHASIGMAD
+	PPC_INS_VSHASIGMAW    = C.PPC_INS_VSHASIGMAW
 	PPC_INS_VSL           = C.PPC_INS_VSL
 	PPC_INS_VSLB          = C.PPC_INS_VSLB
 	PPC_INS_VSLD          = C.PPC_INS_VSLD
 	PPC_INS_VSLDOI        = C.PPC_INS_VSLDOI
 	PPC_INS_VSLH          = C.PPC_INS_VSLH
 	PPC_INS_VSLO          = C.PPC_INS_VSLO
+	PPC_INS_VSLV          = C.PPC_INS_VSLV
 	PPC_INS_VSLW          = C.PPC_INS_VSLW
 	PPC_INS_VSPLTB        = C.PPC_INS_VSPLTB
 	PPC_INS_VSPLTH        = C.PPC_INS_VSPLTH
@@ -1034,8 +1713,12 @@ const (
 	PPC_INS_VSRD          = C.PPC_INS_VSRD
 	PPC_INS_VSRH          = C.PPC_INS_VSRH
 	PPC_INS_VSRO          = C.PPC_INS_VSRO
+	PPC_INS_VSRV          = C.PPC_INS_VSRV
 	PPC_INS_VSRW          = C.PPC_INS_VSRW
+	PPC_INS_VSUBCUQ       = C.PPC_INS_VSUBCUQ
 	PPC_INS_VSUBCUW       = C.PPC_INS_VSUBCUW
+	PPC_INS_VSUBECUQ      = C.PPC_INS_VSUBECUQ
+	PPC_INS_VSUBEUQM      = C.PPC_INS_VSUBEUQM
 	PPC_INS_VSUBFP        = C.PPC_INS_VSUBFP
 	PPC_INS_VSUBSBS       = C.PPC_INS_VSUBSBS
 	PPC_INS_VSUBSHS       = C.PPC_INS_VSUBSHS
@@ -1045,6 +1728,7 @@ const (
 	PPC_INS_VSUBUDM       = C.PPC_INS_VSUBUDM
 	PPC_INS_VSUBUHM       = C.PPC_INS_VSUBUHM
 	PPC_INS_VSUBUHS       = C.PPC_INS_VSUBUHS
+	PPC_INS_VSUBUQM       = C.PPC_INS_VSUBUQM
 	PPC_INS_VSUBUWM       = C.PPC_INS_VSUBUWM
 	PPC_INS_VSUBUWS       = C.PPC_INS_VSUBUWS
 	PPC_INS_VSUM2SWS      = C.PPC_INS_VSUM2SWS
@@ -1055,54 +1739,135 @@ const (
 	PPC_INS_VUPKHPX       = C.PPC_INS_VUPKHPX
 	PPC_INS_VUPKHSB       = C.PPC_INS_VUPKHSB
 	PPC_INS_VUPKHSH       = C.PPC_INS_VUPKHSH
+	PPC_INS_VUPKHSW       = C.PPC_INS_VUPKHSW
 	PPC_INS_VUPKLPX       = C.PPC_INS_VUPKLPX
 	PPC_INS_VUPKLSB       = C.PPC_INS_VUPKLSB
 	PPC_INS_VUPKLSH       = C.PPC_INS_VUPKLSH
+	PPC_INS_VUPKLSW       = C.PPC_INS_VUPKLSW
 	PPC_INS_VXOR          = C.PPC_INS_VXOR
 	PPC_INS_WAIT          = C.PPC_INS_WAIT
+	PPC_INS_WAITIMPL      = C.PPC_INS_WAITIMPL
+	PPC_INS_WAITRSV       = C.PPC_INS_WAITRSV
 	PPC_INS_WRTEE         = C.PPC_INS_WRTEE
 	PPC_INS_WRTEEI        = C.PPC_INS_WRTEEI
+	PPC_INS_XNOP          = C.PPC_INS_XNOP
 	PPC_INS_XOR           = C.PPC_INS_XOR
 	PPC_INS_XORI          = C.PPC_INS_XORI
 	PPC_INS_XORIS         = C.PPC_INS_XORIS
 	PPC_INS_XSABSDP       = C.PPC_INS_XSABSDP
+	PPC_INS_XSABSQP       = C.PPC_INS_XSABSQP
 	PPC_INS_XSADDDP       = C.PPC_INS_XSADDDP
+	PPC_INS_XSADDQP       = C.PPC_INS_XSADDQP
+	PPC_INS_XSADDQPO      = C.PPC_INS_XSADDQPO
+	PPC_INS_XSADDSP       = C.PPC_INS_XSADDSP
+	PPC_INS_XSCMPEQDP     = C.PPC_INS_XSCMPEQDP
+	PPC_INS_XSCMPEXPDP    = C.PPC_INS_XSCMPEXPDP
+	PPC_INS_XSCMPEXPQP    = C.PPC_INS_XSCMPEXPQP
+	PPC_INS_XSCMPGEDP     = C.PPC_INS_XSCMPGEDP
+	PPC_INS_XSCMPGTDP     = C.PPC_INS_XSCMPGTDP
 	PPC_INS_XSCMPODP      = C.PPC_INS_XSCMPODP
+	PPC_INS_XSCMPOQP      = C.PPC_INS_XSCMPOQP
 	PPC_INS_XSCMPUDP      = C.PPC_INS_XSCMPUDP
+	PPC_INS_XSCMPUQP      = C.PPC_INS_XSCMPUQP
 	PPC_INS_XSCPSGNDP     = C.PPC_INS_XSCPSGNDP
+	PPC_INS_XSCPSGNQP     = C.PPC_INS_XSCPSGNQP
+	PPC_INS_XSCVDPHP      = C.PPC_INS_XSCVDPHP
+	PPC_INS_XSCVDPQP      = C.PPC_INS_XSCVDPQP
 	PPC_INS_XSCVDPSP      = C.PPC_INS_XSCVDPSP
+	PPC_INS_XSCVDPSPN     = C.PPC_INS_XSCVDPSPN
 	PPC_INS_XSCVDPSXDS    = C.PPC_INS_XSCVDPSXDS
 	PPC_INS_XSCVDPSXWS    = C.PPC_INS_XSCVDPSXWS
 	PPC_INS_XSCVDPUXDS    = C.PPC_INS_XSCVDPUXDS
 	PPC_INS_XSCVDPUXWS    = C.PPC_INS_XSCVDPUXWS
+	PPC_INS_XSCVHPDP      = C.PPC_INS_XSCVHPDP
+	PPC_INS_XSCVQPDP      = C.PPC_INS_XSCVQPDP
+	PPC_INS_XSCVQPDPO     = C.PPC_INS_XSCVQPDPO
+	PPC_INS_XSCVQPSDZ     = C.PPC_INS_XSCVQPSDZ
+	PPC_INS_XSCVQPSWZ     = C.PPC_INS_XSCVQPSWZ
+	PPC_INS_XSCVQPUDZ     = C.PPC_INS_XSCVQPUDZ
+	PPC_INS_XSCVQPUWZ     = C.PPC_INS_XSCVQPUWZ
+	PPC_INS_XSCVSDQP      = C.PPC_INS_XSCVSDQP
 	PPC_INS_XSCVSPDP      = C.PPC_INS_XSCVSPDP
+	PPC_INS_XSCVSPDPN     = C.PPC_INS_XSCVSPDPN
 	PPC_INS_XSCVSXDDP     = C.PPC_INS_XSCVSXDDP
+	PPC_INS_XSCVSXDSP     = C.PPC_INS_XSCVSXDSP
+	PPC_INS_XSCVUDQP      = C.PPC_INS_XSCVUDQP
 	PPC_INS_XSCVUXDDP     = C.PPC_INS_XSCVUXDDP
+	PPC_INS_XSCVUXDSP     = C.PPC_INS_XSCVUXDSP
 	PPC_INS_XSDIVDP       = C.PPC_INS_XSDIVDP
+	PPC_INS_XSDIVQP       = C.PPC_INS_XSDIVQP
+	PPC_INS_XSDIVQPO      = C.PPC_INS_XSDIVQPO
+	PPC_INS_XSDIVSP       = C.PPC_INS_XSDIVSP
+	PPC_INS_XSIEXPDP      = C.PPC_INS_XSIEXPDP
+	PPC_INS_XSIEXPQP      = C.PPC_INS_XSIEXPQP
 	PPC_INS_XSMADDADP     = C.PPC_INS_XSMADDADP
+	PPC_INS_XSMADDASP     = C.PPC_INS_XSMADDASP
 	PPC_INS_XSMADDMDP     = C.PPC_INS_XSMADDMDP
+	PPC_INS_XSMADDMSP     = C.PPC_INS_XSMADDMSP
+	PPC_INS_XSMADDQP      = C.PPC_INS_XSMADDQP
+	PPC_INS_XSMADDQPO     = C.PPC_INS_XSMADDQPO
+	PPC_INS_XSMAXCDP      = C.PPC_INS_XSMAXCDP
 	PPC_INS_XSMAXDP       = C.PPC_INS_XSMAXDP
+	PPC_INS_XSMAXJDP      = C.PPC_INS_XSMAXJDP
+	PPC_INS_XSMINCDP      = C.PPC_INS_XSMINCDP
 	PPC_INS_XSMINDP       = C.PPC_INS_XSMINDP
+	PPC_INS_XSMINJDP      = C.PPC_INS_XSMINJDP
 	PPC_INS_XSMSUBADP     = C.PPC_INS_XSMSUBADP
+	PPC_INS_XSMSUBASP     = C.PPC_INS_XSMSUBASP
 	PPC_INS_XSMSUBMDP     = C.PPC_INS_XSMSUBMDP
+	PPC_INS_XSMSUBMSP     = C.PPC_INS_XSMSUBMSP
+	PPC_INS_XSMSUBQP      = C.PPC_INS_XSMSUBQP
+	PPC_INS_XSMSUBQPO     = C.PPC_INS_XSMSUBQPO
 	PPC_INS_XSMULDP       = C.PPC_INS_XSMULDP
+	PPC_INS_XSMULQP       = C.PPC_INS_XSMULQP
+	PPC_INS_XSMULQPO      = C.PPC_INS_XSMULQPO
+	PPC_INS_XSMULSP       = C.PPC_INS_XSMULSP
 	PPC_INS_XSNABSDP      = C.PPC_INS_XSNABSDP
+	PPC_INS_XSNABSQP      = C.PPC_INS_XSNABSQP
 	PPC_INS_XSNEGDP       = C.PPC_INS_XSNEGDP
+	PPC_INS_XSNEGQP       = C.PPC_INS_XSNEGQP
 	PPC_INS_XSNMADDADP    = C.PPC_INS_XSNMADDADP
+	PPC_INS_XSNMADDASP    = C.PPC_INS_XSNMADDASP
 	PPC_INS_XSNMADDMDP    = C.PPC_INS_XSNMADDMDP
+	PPC_INS_XSNMADDMSP    = C.PPC_INS_XSNMADDMSP
+	PPC_INS_XSNMADDQP     = C.PPC_INS_XSNMADDQP
+	PPC_INS_XSNMADDQPO    = C.PPC_INS_XSNMADDQPO
 	PPC_INS_XSNMSUBADP    = C.PPC_INS_XSNMSUBADP
+	PPC_INS_XSNMSUBASP    = C.PPC_INS_XSNMSUBASP
 	PPC_INS_XSNMSUBMDP    = C.PPC_INS_XSNMSUBMDP
+	PPC_INS_XSNMSUBMSP    = C.PPC_INS_XSNMSUBMSP
+	PPC_INS_XSNMSUBQP     = C.PPC_INS_XSNMSUBQP
+	PPC_INS_XSNMSUBQPO    = C.PPC_INS_XSNMSUBQPO
 	PPC_INS_XSRDPI        = C.PPC_INS_XSRDPI
 	PPC_INS_XSRDPIC       = C.PPC_INS_XSRDPIC
 	PPC_INS_XSRDPIM       = C.PPC_INS_XSRDPIM
 	PPC_INS_XSRDPIP       = C.PPC_INS_XSRDPIP
 	PPC_INS_XSRDPIZ       = C.PPC_INS_XSRDPIZ
 	PPC_INS_XSREDP        = C.PPC_INS_XSREDP
+	PPC_INS_XSRESP        = C.PPC_INS_XSRESP
+	PPC_INS_XSRQPI        = C.PPC_INS_XSRQPI
+	PPC_INS_XSRQPIX       = C.PPC_INS_XSRQPIX
+	PPC_INS_XSRQPXP       = C.PPC_INS_XSRQPXP
+	PPC_INS_XSRSP         = C.PPC_INS_XSRSP
 	PPC_INS_XSRSQRTEDP    = C.PPC_INS_XSRSQRTEDP
+	PPC_INS_XSRSQRTESP    = C.PPC_INS_XSRSQRTESP
 	PPC_INS_XSSQRTDP      = C.PPC_INS_XSSQRTDP
+	PPC_INS_XSSQRTQP      = C.PPC_INS_XSSQRTQP
+	PPC_INS_XSSQRTQPO     = C.PPC_INS_XSSQRTQPO
+	PPC_INS_XSSQRTSP      = C.PPC_INS_XSSQRTSP
 	PPC_INS_XSSUBDP       = C.PPC_INS_XSSUBDP
+	PPC_INS_XSSUBQP       = C.PPC_INS_XSSUBQP
+	PPC_INS_XSSUBQPO      = C.PPC_INS_XSSUBQPO
+	PPC_INS_XSSUBSP       = C.PPC_INS_XSSUBSP
 	PPC_INS_XSTDIVDP      = C.PPC_INS_XSTDIVDP
 	PPC_INS_XSTSQRTDP     = C.PPC_INS_XSTSQRTDP
+	PPC_INS_XSTSTDCDP     = C.PPC_INS_XSTSTDCDP
+	PPC_INS_XSTSTDCQP     = C.PPC_INS_XSTSTDCQP
+	PPC_INS_XSTSTDCSP     = C.PPC_INS_XSTSTDCSP
+	PPC_INS_XSXEXPDP      = C.PPC_INS_XSXEXPDP
+	PPC_INS_XSXEXPQP      = C.PPC_INS_XSXEXPQP
+	PPC_INS_XSXSIGDP      = C.PPC_INS_XSXSIGDP
+	PPC_INS_XSXSIGQP      = C.PPC_INS_XSXSIGQP
 	PPC_INS_XVABSDP       = C.PPC_INS_XVABSDP
 	PPC_INS_XVABSSP       = C.PPC_INS_XVABSSP
 	PPC_INS_XVADDDP       = C.PPC_INS_XVADDDP
@@ -1120,7 +1885,9 @@ const (
 	PPC_INS_XVCVDPSXWS    = C.PPC_INS_XVCVDPSXWS
 	PPC_INS_XVCVDPUXDS    = C.PPC_INS_XVCVDPUXDS
 	PPC_INS_XVCVDPUXWS    = C.PPC_INS_XVCVDPUXWS
+	PPC_INS_XVCVHPSP      = C.PPC_INS_XVCVHPSP
 	PPC_INS_XVCVSPDP      = C.PPC_INS_XVCVSPDP
+	PPC_INS_XVCVSPHP      = C.PPC_INS_XVCVSPHP
 	PPC_INS_XVCVSPSXDS    = C.PPC_INS_XVCVSPSXDS
 	PPC_INS_XVCVSPSXWS    = C.PPC_INS_XVCVSPSXWS
 	PPC_INS_XVCVSPUXDS    = C.PPC_INS_XVCVSPUXDS
@@ -1135,6 +1902,8 @@ const (
 	PPC_INS_XVCVUXWSP     = C.PPC_INS_XVCVUXWSP
 	PPC_INS_XVDIVDP       = C.PPC_INS_XVDIVDP
 	PPC_INS_XVDIVSP       = C.PPC_INS_XVDIVSP
+	PPC_INS_XVIEXPDP      = C.PPC_INS_XVIEXPDP
+	PPC_INS_XVIEXPSP      = C.PPC_INS_XVIEXPSP
 	PPC_INS_XVMADDADP     = C.PPC_INS_XVMADDADP
 	PPC_INS_XVMADDASP     = C.PPC_INS_XVMADDASP
 	PPC_INS_XVMADDMDP     = C.PPC_INS_XVMADDMDP
@@ -1143,6 +1912,8 @@ const (
 	PPC_INS_XVMAXSP       = C.PPC_INS_XVMAXSP
 	PPC_INS_XVMINDP       = C.PPC_INS_XVMINDP
 	PPC_INS_XVMINSP       = C.PPC_INS_XVMINSP
+	PPC_INS_XVMOVDP       = C.PPC_INS_XVMOVDP
+	PPC_INS_XVMOVSP       = C.PPC_INS_XVMOVSP
 	PPC_INS_XVMSUBADP     = C.PPC_INS_XVMSUBADP
 	PPC_INS_XVMSUBASP     = C.PPC_INS_XVMSUBASP
 	PPC_INS_XVMSUBMDP     = C.PPC_INS_XVMSUBMDP
@@ -1183,6 +1954,18 @@ const (
 	PPC_INS_XVTDIVSP      = C.PPC_INS_XVTDIVSP
 	PPC_INS_XVTSQRTDP     = C.PPC_INS_XVTSQRTDP
 	PPC_INS_XVTSQRTSP     = C.PPC_INS_XVTSQRTSP
+	PPC_INS_XVTSTDCDP     = C.PPC_INS_XVTSTDCDP
+	PPC_INS_XVTSTDCSP     = C.PPC_INS_XVTSTDCSP
+	PPC_INS_XVXEXPDP      = C.PPC_INS_XVXEXPDP
+	PPC_INS_XVXEXPSP      = C.PPC_INS_XVXEXPSP
+	PPC_INS_XVXSIGDP      = C.PPC_INS_XVXSIGDP
+	PPC_INS_XVXSIGSP      = C.PPC_INS_XVXSIGSP
+	PPC_INS_XXBRD         = C.PPC_INS_XXBRD
+	PPC_INS_XXBRH         = C.PPC_INS_XXBRH
+	PPC_INS_XXBRQ         = C.PPC_INS_XXBRQ
+	PPC_INS_XXBRW         = C.PPC_INS_XXBRW
+	PPC_INS_XXEXTRACTUW   = C.PPC_INS_XXEXTRACTUW
+	PPC_INS_XXINSERTW     = C.PPC_INS_XXINSERTW
 	PPC_INS_XXLAND        = C.PPC_INS_XXLAND
 	PPC_INS_XXLANDC       = C.PPC_INS_XXLANDC
 	PPC_INS_XXLEQV        = C.PPC_INS_XXLEQV
@@ -1191,191 +1974,57 @@ const (
 	PPC_INS_XXLOR         = C.PPC_INS_XXLOR
 	PPC_INS_XXLORC        = C.PPC_INS_XXLORC
 	PPC_INS_XXLXOR        = C.PPC_INS_XXLXOR
+	PPC_INS_XXMRGHD       = C.PPC_INS_XXMRGHD
 	PPC_INS_XXMRGHW       = C.PPC_INS_XXMRGHW
+	PPC_INS_XXMRGLD       = C.PPC_INS_XXMRGLD
 	PPC_INS_XXMRGLW       = C.PPC_INS_XXMRGLW
+	PPC_INS_XXPERM        = C.PPC_INS_XXPERM
 	PPC_INS_XXPERMDI      = C.PPC_INS_XXPERMDI
+	PPC_INS_XXPERMR       = C.PPC_INS_XXPERMR
 	PPC_INS_XXSEL         = C.PPC_INS_XXSEL
 	PPC_INS_XXSLDWI       = C.PPC_INS_XXSLDWI
-	PPC_INS_XXSPLTW       = C.PPC_INS_XXSPLTW
-	PPC_INS_BCA           = C.PPC_INS_BCA
-	PPC_INS_BCLA          = C.PPC_INS_BCLA
-	PPC_INS_SLWI          = C.PPC_INS_SLWI
-	PPC_INS_SRWI          = C.PPC_INS_SRWI
-	PPC_INS_SLDI          = C.PPC_INS_SLDI
-	PPC_INS_BTA           = C.PPC_INS_BTA
-	PPC_INS_CRSET         = C.PPC_INS_CRSET
-	PPC_INS_CRNOT         = C.PPC_INS_CRNOT
-	PPC_INS_CRMOVE        = C.PPC_INS_CRMOVE
-	PPC_INS_CRCLR         = C.PPC_INS_CRCLR
-	PPC_INS_MFBR0         = C.PPC_INS_MFBR0
-	PPC_INS_MFBR1         = C.PPC_INS_MFBR1
-	PPC_INS_MFBR2         = C.PPC_INS_MFBR2
-	PPC_INS_MFBR3         = C.PPC_INS_MFBR3
-	PPC_INS_MFBR4         = C.PPC_INS_MFBR4
-	PPC_INS_MFBR5         = C.PPC_INS_MFBR5
-	PPC_INS_MFBR6         = C.PPC_INS_MFBR6
-	PPC_INS_MFBR7         = C.PPC_INS_MFBR7
-	PPC_INS_MFXER         = C.PPC_INS_MFXER
-	PPC_INS_MFRTCU        = C.PPC_INS_MFRTCU
-	PPC_INS_MFRTCL        = C.PPC_INS_MFRTCL
-	PPC_INS_MFDSCR        = C.PPC_INS_MFDSCR
-	PPC_INS_MFDSISR       = C.PPC_INS_MFDSISR
-	PPC_INS_MFDAR         = C.PPC_INS_MFDAR
-	PPC_INS_MFSRR2        = C.PPC_INS_MFSRR2
-	PPC_INS_MFSRR3        = C.PPC_INS_MFSRR3
-	PPC_INS_MFCFAR        = C.PPC_INS_MFCFAR
-	PPC_INS_MFAMR         = C.PPC_INS_MFAMR
-	PPC_INS_MFPID         = C.PPC_INS_MFPID
-	PPC_INS_MFTBLO        = C.PPC_INS_MFTBLO
-	PPC_INS_MFTBHI        = C.PPC_INS_MFTBHI
-	PPC_INS_MFDBATU       = C.PPC_INS_MFDBATU
-	PPC_INS_MFDBATL       = C.PPC_INS_MFDBATL
-	PPC_INS_MFIBATU       = C.PPC_INS_MFIBATU
-	PPC_INS_MFIBATL       = C.PPC_INS_MFIBATL
-	PPC_INS_MFDCCR        = C.PPC_INS_MFDCCR
-	PPC_INS_MFICCR        = C.PPC_INS_MFICCR
-	PPC_INS_MFDEAR        = C.PPC_INS_MFDEAR
-	PPC_INS_MFESR         = C.PPC_INS_MFESR
-	PPC_INS_MFSPEFSCR     = C.PPC_INS_MFSPEFSCR
-	PPC_INS_MFTCR         = C.PPC_INS_MFTCR
-	PPC_INS_MFASR         = C.PPC_INS_MFASR
-	PPC_INS_MFPVR         = C.PPC_INS_MFPVR
-	PPC_INS_MFTBU         = C.PPC_INS_MFTBU
-	PPC_INS_MTCR          = C.PPC_INS_MTCR
-	PPC_INS_MTBR0         = C.PPC_INS_MTBR0
-	PPC_INS_MTBR1         = C.PPC_INS_MTBR1
-	PPC_INS_MTBR2         = C.PPC_INS_MTBR2
-	PPC_INS_MTBR3         = C.PPC_INS_MTBR3
-	PPC_INS_MTBR4         = C.PPC_INS_MTBR4
-	PPC_INS_MTBR5         = C.PPC_INS_MTBR5
-	PPC_INS_MTBR6         = C.PPC_INS_MTBR6
-	PPC_INS_MTBR7         = C.PPC_INS_MTBR7
-	PPC_INS_MTXER         = C.PPC_INS_MTXER
-	PPC_INS_MTDSCR        = C.PPC_INS_MTDSCR
-	PPC_INS_MTDSISR       = C.PPC_INS_MTDSISR
-	PPC_INS_MTDAR         = C.PPC_INS_MTDAR
-	PPC_INS_MTSRR2        = C.PPC_INS_MTSRR2
-	PPC_INS_MTSRR3        = C.PPC_INS_MTSRR3
-	PPC_INS_MTCFAR        = C.PPC_INS_MTCFAR
-	PPC_INS_MTAMR         = C.PPC_INS_MTAMR
-	PPC_INS_MTPID         = C.PPC_INS_MTPID
-	PPC_INS_MTTBL         = C.PPC_INS_MTTBL
-	PPC_INS_MTTBU         = C.PPC_INS_MTTBU
-	PPC_INS_MTTBLO        = C.PPC_INS_MTTBLO
-	PPC_INS_MTTBHI        = C.PPC_INS_MTTBHI
-	PPC_INS_MTDBATU       = C.PPC_INS_MTDBATU
-	PPC_INS_MTDBATL       = C.PPC_INS_MTDBATL
-	PPC_INS_MTIBATU       = C.PPC_INS_MTIBATU
-	PPC_INS_MTIBATL       = C.PPC_INS_MTIBATL
-	PPC_INS_MTDCCR        = C.PPC_INS_MTDCCR
-	PPC_INS_MTICCR        = C.PPC_INS_MTICCR
-	PPC_INS_MTDEAR        = C.PPC_INS_MTDEAR
-	PPC_INS_MTESR         = C.PPC_INS_MTESR
-	PPC_INS_MTSPEFSCR     = C.PPC_INS_MTSPEFSCR
-	PPC_INS_MTTCR         = C.PPC_INS_MTTCR
-	PPC_INS_NOT           = C.PPC_INS_NOT
-	PPC_INS_MR            = C.PPC_INS_MR
-	PPC_INS_ROTLD         = C.PPC_INS_ROTLD
-	PPC_INS_ROTLDI        = C.PPC_INS_ROTLDI
-	PPC_INS_CLRLDI        = C.PPC_INS_CLRLDI
-	PPC_INS_ROTLWI        = C.PPC_INS_ROTLWI
-	PPC_INS_CLRLWI        = C.PPC_INS_CLRLWI
-	PPC_INS_ROTLW         = C.PPC_INS_ROTLW
-	PPC_INS_SUB           = C.PPC_INS_SUB
-	PPC_INS_SUBC          = C.PPC_INS_SUBC
-	PPC_INS_LWSYNC        = C.PPC_INS_LWSYNC
-	PPC_INS_PTESYNC       = C.PPC_INS_PTESYNC
-	PPC_INS_TDLT          = C.PPC_INS_TDLT
-	PPC_INS_TDEQ          = C.PPC_INS_TDEQ
-	PPC_INS_TDGT          = C.PPC_INS_TDGT
-	PPC_INS_TDNE          = C.PPC_INS_TDNE
-	PPC_INS_TDLLT         = C.PPC_INS_TDLLT
-	PPC_INS_TDLGT         = C.PPC_INS_TDLGT
-	PPC_INS_TDU           = C.PPC_INS_TDU
-	PPC_INS_TDLTI         = C.PPC_INS_TDLTI
-	PPC_INS_TDEQI         = C.PPC_INS_TDEQI
-	PPC_INS_TDGTI         = C.PPC_INS_TDGTI
-	PPC_INS_TDNEI         = C.PPC_INS_TDNEI
-	PPC_INS_TDLLTI        = C.PPC_INS_TDLLTI
-	PPC_INS_TDLGTI        = C.PPC_INS_TDLGTI
-	PPC_INS_TDUI          = C.PPC_INS_TDUI
-	PPC_INS_TLBREHI       = C.PPC_INS_TLBREHI
-	PPC_INS_TLBRELO       = C.PPC_INS_TLBRELO
-	PPC_INS_TLBWEHI       = C.PPC_INS_TLBWEHI
-	PPC_INS_TLBWELO       = C.PPC_INS_TLBWELO
-	PPC_INS_TWLT          = C.PPC_INS_TWLT
-	PPC_INS_TWEQ          = C.PPC_INS_TWEQ
-	PPC_INS_TWGT          = C.PPC_INS_TWGT
-	PPC_INS_TWNE          = C.PPC_INS_TWNE
-	PPC_INS_TWLLT         = C.PPC_INS_TWLLT
-	PPC_INS_TWLGT         = C.PPC_INS_TWLGT
-	PPC_INS_TWU           = C.PPC_INS_TWU
-	PPC_INS_TWLTI         = C.PPC_INS_TWLTI
-	PPC_INS_TWEQI         = C.PPC_INS_TWEQI
-	PPC_INS_TWGTI         = C.PPC_INS_TWGTI
-	PPC_INS_TWNEI         = C.PPC_INS_TWNEI
-	PPC_INS_TWLLTI        = C.PPC_INS_TWLLTI
-	PPC_INS_TWLGTI        = C.PPC_INS_TWLGTI
-	PPC_INS_TWUI          = C.PPC_INS_TWUI
-	PPC_INS_WAITRSV       = C.PPC_INS_WAITRSV
-	PPC_INS_WAITIMPL      = C.PPC_INS_WAITIMPL
-	PPC_INS_XNOP          = C.PPC_INS_XNOP
-	PPC_INS_XVMOVDP       = C.PPC_INS_XVMOVDP
-	PPC_INS_XVMOVSP       = C.PPC_INS_XVMOVSP
 	PPC_INS_XXSPLTD       = C.PPC_INS_XXSPLTD
-	PPC_INS_XXMRGHD       = C.PPC_INS_XXMRGHD
-	PPC_INS_XXMRGLD       = C.PPC_INS_XXMRGLD
+	PPC_INS_XXSPLTIB      = C.PPC_INS_XXSPLTIB
+	PPC_INS_XXSPLTW       = C.PPC_INS_XXSPLTW
 	PPC_INS_XXSWAPD       = C.PPC_INS_XXSWAPD
-	PPC_INS_BT            = C.PPC_INS_BT
-	PPC_INS_BF            = C.PPC_INS_BF
-	PPC_INS_BDNZT         = C.PPC_INS_BDNZT
-	PPC_INS_BDNZF         = C.PPC_INS_BDNZF
-	PPC_INS_BDZF          = C.PPC_INS_BDZF
-	PPC_INS_BDZT          = C.PPC_INS_BDZT
-	PPC_INS_BFA           = C.PPC_INS_BFA
-	PPC_INS_BDNZTA        = C.PPC_INS_BDNZTA
-	PPC_INS_BDNZFA        = C.PPC_INS_BDNZFA
-	PPC_INS_BDZTA         = C.PPC_INS_BDZTA
-	PPC_INS_BDZFA         = C.PPC_INS_BDZFA
-	PPC_INS_BTCTR         = C.PPC_INS_BTCTR
-	PPC_INS_BFCTR         = C.PPC_INS_BFCTR
-	PPC_INS_BTCTRL        = C.PPC_INS_BTCTRL
-	PPC_INS_BFCTRL        = C.PPC_INS_BFCTRL
-	PPC_INS_BTL           = C.PPC_INS_BTL
-	PPC_INS_BFL           = C.PPC_INS_BFL
-	PPC_INS_BDNZTL        = C.PPC_INS_BDNZTL
-	PPC_INS_BDNZFL        = C.PPC_INS_BDNZFL
-	PPC_INS_BDZTL         = C.PPC_INS_BDZTL
-	PPC_INS_BDZFL         = C.PPC_INS_BDZFL
-	PPC_INS_BTLA          = C.PPC_INS_BTLA
-	PPC_INS_BFLA          = C.PPC_INS_BFLA
-	PPC_INS_BDNZTLA       = C.PPC_INS_BDNZTLA
-	PPC_INS_BDNZFLA       = C.PPC_INS_BDNZFLA
-	PPC_INS_BDZTLA        = C.PPC_INS_BDZTLA
-	PPC_INS_BDZFLA        = C.PPC_INS_BDZFLA
-	PPC_INS_BTLR          = C.PPC_INS_BTLR
-	PPC_INS_BFLR          = C.PPC_INS_BFLR
-	PPC_INS_BDNZTLR       = C.PPC_INS_BDNZTLR
-	PPC_INS_BDZTLR        = C.PPC_INS_BDZTLR
-	PPC_INS_BDZFLR        = C.PPC_INS_BDZFLR
-	PPC_INS_BTLRL         = C.PPC_INS_BTLRL
-	PPC_INS_BFLRL         = C.PPC_INS_BFLRL
-	PPC_INS_BDNZTLRL      = C.PPC_INS_BDNZTLRL
-	PPC_INS_BDNZFLRL      = C.PPC_INS_BDNZFLRL
-	PPC_INS_BDZTLRL       = C.PPC_INS_BDZTLRL
-	PPC_INS_BDZFLRL       = C.PPC_INS_BDZFLRL
-	PPC_INS_QVFAND        = C.PPC_INS_QVFAND
-	PPC_INS_QVFCLR        = C.PPC_INS_QVFCLR
-	PPC_INS_QVFANDC       = C.PPC_INS_QVFANDC
-	PPC_INS_QVFCTFB       = C.PPC_INS_QVFCTFB
-	PPC_INS_QVFXOR        = C.PPC_INS_QVFXOR
-	PPC_INS_QVFOR         = C.PPC_INS_QVFOR
-	PPC_INS_QVFNOR        = C.PPC_INS_QVFNOR
-	PPC_INS_QVFEQU        = C.PPC_INS_QVFEQU
-	PPC_INS_QVFNOT        = C.PPC_INS_QVFNOT
-	PPC_INS_QVFORC        = C.PPC_INS_QVFORC
-	PPC_INS_QVFNAND       = C.PPC_INS_QVFNAND
-	PPC_INS_QVFSET        = C.PPC_INS_QVFSET
+	PPC_INS_DCBZ_L        = C.PPC_INS_DCBZ_L
+	PPC_INS_PSQ_L         = C.PPC_INS_PSQ_L
+	PPC_INS_PSQ_LU        = C.PPC_INS_PSQ_LU
+	PPC_INS_PSQ_LUX       = C.PPC_INS_PSQ_LUX
+	PPC_INS_PSQ_LX        = C.PPC_INS_PSQ_LX
+	PPC_INS_PSQ_ST        = C.PPC_INS_PSQ_ST
+	PPC_INS_PSQ_STU       = C.PPC_INS_PSQ_STU
+	PPC_INS_PSQ_STUX      = C.PPC_INS_PSQ_STUX
+	PPC_INS_PSQ_STX       = C.PPC_INS_PSQ_STX
+	PPC_INS_PS_ABS        = C.PPC_INS_PS_ABS
+	PPC_INS_PS_ADD        = C.PPC_INS_PS_ADD
+	PPC_INS_PS_CMPO0      = C.PPC_INS_PS_CMPO0
+	PPC_INS_PS_CMPO1      = C.PPC_INS_PS_CMPO1
+	PPC_INS_PS_CMPU0      = C.PPC_INS_PS_CMPU0
+	PPC_INS_PS_CMPU1      = C.PPC_INS_PS_CMPU1
+	PPC_INS_PS_DIV        = C.PPC_INS_PS_DIV
+	PPC_INS_PS_MADD       = C.PPC_INS_PS_MADD
+	PPC_INS_PS_MADDS0     = C.PPC_INS_PS_MADDS0
+	PPC_INS_PS_MADDS1     = C.PPC_INS_PS_MADDS1
+	PPC_INS_PS_MERGE00    = C.PPC_INS_PS_MERGE00
+	PPC_INS_PS_MERGE01    = C.PPC_INS_PS_MERGE01
+	PPC_INS_PS_MERGE10    = C.PPC_INS_PS_MERGE10
+	PPC_INS_PS_MERGE11    = C.PPC_INS_PS_MERGE11
+	PPC_INS_PS_MR         = C.PPC_INS_PS_MR
+	PPC_INS_PS_MSUB       = C.PPC_INS_PS_MSUB
+	PPC_INS_PS_MUL        = C.PPC_INS_PS_MUL
+	PPC_INS_PS_MULS0      = C.PPC_INS_PS_MULS0
+	PPC_INS_PS_MULS1      = C.PPC_INS_PS_MULS1
+	PPC_INS_PS_NABS       = C.PPC_INS_PS_NABS
+	PPC_INS_PS_NEG        = C.PPC_INS_PS_NEG
+	PPC_INS_PS_NMADD      = C.PPC_INS_PS_NMADD
+	PPC_INS_PS_NMSUB      = C.PPC_INS_PS_NMSUB
+	PPC_INS_PS_RES        = C.PPC_INS_PS_RES
+	PPC_INS_PS_RSQRTE     = C.PPC_INS_PS_RSQRTE
+	PPC_INS_PS_SEL        = C.PPC_INS_PS_SEL
+	PPC_INS_PS_SUB        = C.PPC_INS_PS_SUB
+	PPC_INS_PS_SUM0       = C.PPC_INS_PS_SUM0
+	PPC_INS_PS_SUM1       = C.PPC_INS_PS_SUM1
 	PPC_INS_ENDING        = C.PPC_INS_ENDING
 )
 
@@ -1396,5 +2045,6 @@ const (
 	PPC_GRP_P8ALTIVEC = C.PPC_GRP_P8ALTIVEC
 	PPC_GRP_P8VECTOR  = C.PPC_GRP_P8VECTOR
 	PPC_GRP_QPX       = C.PPC_GRP_QPX
+	PPC_GRP_PS        = C.PPC_GRP_PS
 	PPC_GRP_ENDING    = C.PPC_GRP_ENDING
 )
