@@ -16,7 +16,7 @@ func TestVersion(t *testing.T) {
 	if c, err := New(0, 0); err == nil {
 		maj, min := c.Version()
 		if maj == checks.Maj() && min == checks.Min() {
-			t.Logf("Libary version %v.%v, OK.", maj, min)
+			t.Logf("Library version %v.%v, OK.", maj, min)
 			t.Logf("CAPSTONE_DIET: %v", dietMode)
 		} else {
 			t.Errorf(
